Calculate correlation between technical replicates

Calculate and return list of correlation between paired technical replicates in a dataset

Usage

calcReplicateCorr(dge, group, corr.thresh = 0.8, return = "all")

Arguments

dge

DGEList object containing technical replicates to be compared

group

grouping variable from metadata containing technical replicate information

corr.thresh

threshold distinguishing good vs bad correlation between technical replicates

return

which values to return. One of "good", "bad", "all"

Value

Returns a plot of the read counts per barcode (row) in a data frame

Examples

data(test.dge)
calcReplicateCorr(test.dge, group = "group")
#>   1_BM_High_dose  1_SPL_High_dose  10_BM_High_dose 10_SPL_High_dose 
#>        0.9990491        0.8948741        0.9998575        0.9997446 
#>    11_BM_Vehicle   11_SPL_Vehicle    12_BM_Vehicle   12_SPL_Vehicle 
#>        0.9988328        0.9990118        0.9992805        0.9996900 
#>   13_BM_Low_dose  13_SPL_Low_dose   14_BM_Low_dose  14_SPL_Low_dose 
#>        0.9999569        0.9999258        0.9999779        0.9996698 
#>   15_BM_Low_dose  15_SPL_Low_dose   16_BM_Low_dose  16_SPL_Low_dose 
#>        0.9998939        0.9999679        0.9998965        0.9996074 
#>  17_BM_High_dose 17_SPL_High_dose    18_BM_Vehicle   18_SPL_Vehicle 
#>        0.9998268        0.9996864        0.9995628        0.9993980 
#>   2_BM_High_dose  2_SPL_High_dose   3_BM_High_dose  3_SPL_High_dose 
#>        0.9975029        0.9973275        0.9997352        0.9998276 
#>     4_BM_Vehicle    4_SPL_Vehicle    5_BM_Low_dose   5_SPL_Low_dose 
#>        0.9975620        0.9986015        0.9999389        0.9978161 
#>   6_BM_High_dose  6_SPL_High_dose     7_BM_Vehicle    7_SPL_Vehicle 
#>        0.9994985        0.9997586        0.9997088        0.9972775 
#>     8_BM_Vehicle    8_SPL_Vehicle    9_BM_Low_dose   9_SPL_Low_dose 
#>        0.9992879        0.9989773        0.9997057        0.9993462 
#>               T0              T0b 
#>        0.9983838        0.9963650