Collapse technical replicates

Collapse technical replicates in a DGEList object by mean or sum. Modified from collapseReplicates function from DESeq2 package to accept DGEList objects and to allow collapsing replicated by mean or sum.

Usage

collapseReplicates(
  object,
  groupby,
  run,
  renameCols = TRUE,
  show_reps = TRUE,
  by = "mean"
)

Arguments

object

DGEList object containing raw or normalised barcode counts with replicate grouping info in object metadata

groupby

a character vector of containing grouping information. Must be as long as the columns of object. Can pass a metadata column from object.

run

optional, the names of each unique column in object. if provided, a new column runsCollapsed will be added to the colData which pastes together the names of run

renameCols

boolean. whether to rename the columns of the returned object using the levels of the grouping factor

show_reps

boolean. print replicate column names to console?

by

collapse replicates by mean or sum

Value

Returns a dataframe of normalised counts for a sample

Examples

data(test.dge)
collapseReplicates(test.dge, groupby = test.dge$samples$group, by = "mean")
#> $`10_High_dose`
#> [1] 3 4
#> 
#> $`11_Vehicle`
#> [1] 5 6
#> 
#> $`12_Vehicle`
#> [1] 7 8
#> 
#> $`13_Low_dose`
#> [1]  9 10
#> 
#> $`14_Low_dose`
#> [1] 11 12
#> 
#> $`15_Low_dose`
#> [1] 13 14
#> 
#> $`16_Low_dose`
#> [1] 15 16
#> 
#> $`17_High_dose`
#> [1] 17 18
#> 
#> $`18_Vehicle`
#> [1] 19 20
#> 
#> $`1_High_dose`
#> [1] 21 22
#> 
#> $`2_High_dose`
#> [1] 23 24
#> 
#> $`3_High_dose`
#> [1] 25 26
#> 
#> $`4_Vehicle`
#> [1] 27 28
#> 
#> $`5_Low_dose`
#> [1] 29 30
#> 
#> $`6_High_dose`
#> [1] 31 32
#> 
#> $`7_Vehicle`
#> [1] 33 34
#> 
#> $`8_Vehicle`
#> [1] 35 36
#> 
#> $`9_Low_dose`
#> [1] 37 38
#> 
#> $T0
#> [1] 1 2
#> 
#> An object of class "DGEList"
#> $samples
#>              Sample Experiment        Group PCR_Replicate Treatment
#> 10_High_dose  S10-1    test_01 10_High_dose             1 High_dose
#> 11_Vehicle    S11-1    test_01   11_Vehicle             1   Vehicle
#> 12_Vehicle    S12-1    test_01   12_Vehicle             1   Vehicle
#> 13_Low_dose   S13-1    test_01  13_Low_dose             1  Low_dose
#> 14_Low_dose   S14-1    test_01  14_Low_dose             1  Low_dose
#>                     group lib.size norm.factors
#> 10_High_dose 10_High_dose  4114186            1
#> 11_Vehicle     11_Vehicle  2907500            1
#> 12_Vehicle     12_Vehicle  4202337            1
#> 13_Low_dose   13_Low_dose  4513559            1
#> 14_Low_dose   14_Low_dose  4503490            1
#> 14 more rows ...
#> 
#> $counts
#>         Samples
#> Tags     10_High_dose 11_Vehicle 12_Vehicle 13_Low_dose 14_Low_dose 15_Low_dose
#>   BC_1              0          0          0           0           0           0
#>   BC_13             0          0          0           0           0           0
#>   BC_99             0          0          0           0           0           0
#>   BC_120            0          0          0           0           0           0
#>   BC_351            0          0          0           0           0           0
#>         Samples
#> Tags     16_Low_dose 17_High_dose 18_Vehicle 1_High_dose 2_High_dose
#>   BC_1             0            0          0           0           0
#>   BC_13            0            0          0           0           0
#>   BC_99            0            0          0         155           0
#>   BC_120           0            0          0           0           0
#>   BC_351           0            0          0           0           0
#>         Samples
#> Tags     3_High_dose 4_Vehicle 5_Low_dose 6_High_dose 7_Vehicle 8_Vehicle
#>   BC_1             0         0          0           0         0         0
#>   BC_13            0         0          0           0         0         0
#>   BC_99            0         0          0           0         0         0
#>   BC_120           0         0          0           0         0         0
#>   BC_351           0         0          0           0         0         0
#>         Samples
#> Tags     9_Low_dose   T0
#>   BC_1            0  127
#>   BC_13           0 1146
#>   BC_99           0 1354
#>   BC_120          0  234
#>   BC_351          0    0
#> 1634 more rows ...
#>