Barcode abundance comparison

Takes a dataframe of barcode counts, computes the mean abundance of each barcode for two specific conditions, then do a line plot for both conditions

Usage

compareAbundance(counts, condition, condition_names)

Arguments

counts

dataframe containing raw counts of barcodes

condition

sample condition of interest

condition_names

names of possible conditions in experiment

Value

Returns a text file

Details

compareAbundance

Compare barcode abundance between 2 conditions using EdgeR

Examples

data(test.dge)
compareAbundance(test.dge, test.dge$samples$group, c("T0","T0b"))
#> Disp = 0.4546 , BCV = 0.6743 
#>                      logFC  logCPM    PValue
#> BFP_Barcode_65784  -12.129  6.5167 1.824e-11
#> BFP_Barcode_74018   11.531  5.9196 1.963e-10
#> BFP_Barcode_14630  -11.440  5.8353 2.250e-09
#> BFP_Barcode_149175 -10.649  5.0572 3.329e-09
#> BFP_Barcode_152974 -10.726  5.1322 3.728e-09
#> BFP_Barcode_236573 -10.640  5.0483 4.035e-09
#> BFP_Barcode_493429 -10.510  4.9209 4.900e-09
#> BFP_Barcode_5919    10.753  5.1501 5.028e-09
#> BFP_Barcode_326988  10.926  5.3207 9.380e-09
#> BFP_Barcode_124495  10.207  4.6153 2.402e-08
#> BFP_Barcode_334190 -10.032  4.4568 2.408e-08
#> BFP_Barcode_196294  10.368  4.7717 2.819e-08
#> BFP_Barcode_472825 -10.112  4.5343 3.723e-08
#> BFP_Barcode_139720  -9.921  4.3494 4.045e-08
#> BFP_Barcode_294255  -9.893  4.3221 4.833e-08
#> BFP_Barcode_388933  -9.794  4.2268 6.766e-08
#> BFP_Barcode_90987   -9.917  4.3456 7.332e-08
#> BFP_Barcode_523732   9.931  4.3445 7.373e-08
#> BFP_Barcode_18485    9.791  4.2106 9.015e-08
#> BFP_Barcode_7131    -9.665  4.1028 1.096e-07
#> BFP_Barcode_280253 -10.102  4.5247 1.196e-07
#> BFP_Barcode_218871   9.489  3.9174 1.515e-07
#> BFP_Barcode_420973 -10.032  4.4566 1.587e-07
#> BFP_Barcode_5924    -9.422  3.8707 1.879e-07
#> BFP_Barcode_604725  -9.432  3.8806 2.352e-07
#> BFP_Barcode_95233   -9.479  3.9250 2.401e-07
#> BFP_Barcode_373069  -9.574  4.0161 2.521e-07
#> BFP_Barcode_2704    -9.904  4.3323 2.544e-07
#> BFP_Barcode_58355    9.810  4.2297 2.697e-07
#> BFP_Barcode_280615  -9.268  3.7245 3.283e-07
#> BFP_Barcode_189714  -9.227  3.6863 3.753e-07
#> BFP_Barcode_308856   9.222  3.6614 3.759e-07
#> BFP_Barcode_145474  -9.422  3.8714 5.004e-07
#> BFP_Barcode_55777   10.108  4.5158 5.938e-07
#> BFP_Barcode_21416   -9.050  3.5196 6.358e-07
#> BFP_Barcode_57262   10.384  4.7867 6.440e-07
#> BFP_Barcode_280772   7.476  4.9135 7.278e-07
#> BFP_Barcode_200918  -9.178  3.6400 8.572e-07
#> BFP_Barcode_317400  -9.089  3.5554 1.140e-06
#> BFP_Barcode_74771   -8.942  3.4185 1.178e-06
#> BFP_Barcode_176499  -9.148  3.6116 1.316e-06
#> BFP_Barcode_33582   -9.164  3.6273 1.360e-06
#> BFP_Barcode_272416  -9.062  3.5303 1.479e-06
#> BFP_Barcode_45050   -8.805  3.2908 1.485e-06
#> BFP_Barcode_290540   7.545  4.1395 1.542e-06
#> BFP_Barcode_23394   -9.670  4.1075 1.573e-06
#> BFP_Barcode_37899   -8.716  3.2083 1.900e-06
#> BFP_Barcode_24589   -8.725  3.2167 2.041e-06
#> BFP_Barcode_142905   7.245  3.8548 2.186e-06
#> BFP_Barcode_79183   -8.816  3.3018 2.340e-06
#> BFP_Barcode_167030  -8.773  3.2615 2.371e-06
#> BFP_Barcode_14563   -8.621  3.1207 2.705e-06
#> BFP_Barcode_289757  -8.966  3.4410 3.137e-06
#> BFP_Barcode_371709   8.593  3.0649 3.293e-06
#> BFP_Barcode_4972    -9.443  3.8919 3.365e-06
#> BFP_Barcode_341320  -8.508  3.0170 3.744e-06
#> BFP_Barcode_183591   9.051  3.4951 3.948e-06
#> BFP_Barcode_125969  -8.894  3.3738 3.961e-06
#> BFP_Barcode_119927  -8.622  3.1214 4.620e-06
#> BFP_Barcode_30420   -8.479  2.9901 5.190e-06
#> BFP_Barcode_49629    7.582  4.1787 5.524e-06
#> BFP_Barcode_482248   6.021  4.6425 5.650e-06
#> BFP_Barcode_203576   5.919  4.2527 5.830e-06
#> BFP_Barcode_486425   7.351  3.9757 6.464e-06
#> BFP_Barcode_373157  -8.389  2.9087 6.652e-06
#> BFP_Barcode_113707   8.390  2.8781 7.040e-06
#> BFP_Barcode_165608  -8.854  3.3353 7.858e-06
#> BFP_Barcode_127186  -8.363  2.8846 8.980e-06
#> BFP_Barcode_501627  -8.418  2.9356 9.232e-06
#> BFP_Barcode_206327   6.234  5.2996 9.363e-06
#> BFP_Barcode_39151    9.064  3.5137 1.021e-05
#> BFP_Barcode_55938   -8.646  3.1433 1.055e-05
#> BFP_Barcode_79511   -8.416  2.9338 1.122e-05
#> BFP_Barcode_21912   -8.247  2.7803 1.162e-05
#> BFP_Barcode_128877  -5.049  6.2187 1.260e-05
#> BFP_Barcode_137154  -8.249  2.7832 1.359e-05
#> BFP_Barcode_116727   8.347  2.8343 1.607e-05
#> BFP_Barcode_398459  -5.947  5.0557 1.734e-05
#> BFP_Barcode_212460  -9.112  3.5785 1.800e-05
#> BFP_Barcode_144876   9.531  3.9620 1.846e-05
#> BFP_Barcode_231990  -8.126  2.6721 1.910e-05
#> BFP_Barcode_74133   -8.053  2.6080 2.622e-05
#> BFP_Barcode_97785   -8.226  2.7611 2.821e-05
#> BFP_Barcode_445631   7.999  2.5223 2.868e-05
#> BFP_Barcode_16372   -8.087  2.6382 3.027e-05
#> BFP_Barcode_31178   -8.682  3.1760 3.520e-05
#> BFP_Barcode_164274  -7.930  2.4982 3.587e-05
#> BFP_Barcode_226861  -8.365  2.8861 3.762e-05
#> BFP_Barcode_47092   -5.427  3.2606 3.821e-05
#> BFP_Barcode_113591   7.940  2.4700 4.049e-05
#> BFP_Barcode_26702   -7.791  2.3769 4.120e-05
#> BFP_Barcode_196731  -8.250  2.7819 4.796e-05
#> BFP_Barcode_71582    7.711  2.2629 4.938e-05
#> BFP_Barcode_122842   4.471  5.9399 5.122e-05
#> BFP_Barcode_314613  -8.045  2.5992 6.209e-05
#> BFP_Barcode_282725  -7.607  2.2189 6.432e-05
#> BFP_Barcode_517468   4.632  6.4515 6.569e-05
#> BFP_Barcode_291556   5.271  4.7105 6.940e-05
#> BFP_Barcode_218710  -7.598  2.2117 7.168e-05
#> BFP_Barcode_200708  -7.596  2.2102 7.490e-05
#> BFP_Barcode_43888    4.237  7.9013 7.544e-05
#> BFP_Barcode_556209  -4.880  3.6439 9.442e-05
#> BFP_Barcode_9376    -8.257  2.7879 1.105e-04
#> BFP_Barcode_615275  -8.126  2.6739 1.145e-04
#> BFP_Barcode_194917  -7.430  2.0685 1.445e-04
#> BFP_Barcode_184509  -8.123  2.6717 1.659e-04
#> BFP_Barcode_141881  -7.333  1.9879 1.693e-04
#> BFP_Barcode_353227   4.229  5.4173 2.062e-04
#> BFP_Barcode_26118   -4.852  2.7482 2.523e-04
#> BFP_Barcode_21051    8.149  2.6482 2.770e-04
#> BFP_Barcode_4079     3.712  5.2015 3.409e-04
#> BFP_Barcode_165146  -7.716  2.3147 4.050e-04
#> BFP_Barcode_257559  -4.143  4.7850 4.179e-04
#> BFP_Barcode_35792   -6.965  1.6899 4.216e-04
#> BFP_Barcode_548028   4.634  2.7418 5.621e-04
#> BFP_Barcode_211257   6.970  1.6318 6.174e-04
#> BFP_Barcode_16921    3.609  6.6101 6.857e-04
#> BFP_Barcode_415460   3.624  4.9002 6.975e-04
#> BFP_Barcode_262018   3.527  4.5845 7.387e-04
#> BFP_Barcode_239000   4.462  4.0221 7.572e-04
#> BFP_Barcode_141180  -6.697  1.4812 8.570e-04
#> BFP_Barcode_221691   3.267 11.2847 1.016e-03
#> BFP_Barcode_112299   3.731  6.3619 1.127e-03
#> BFP_Barcode_398530  -3.270  6.1960 1.319e-03
#> BFP_Barcode_7249     3.508  8.3469 1.326e-03
#> BFP_Barcode_33738   -6.533  1.3581 1.343e-03
#> BFP_Barcode_627295  -3.233  4.6751 1.381e-03
#> BFP_Barcode_953     -5.493  3.4226 1.980e-03
#> BFP_Barcode_4469     3.395  4.7472 2.156e-03
#> BFP_Barcode_345088  -2.979  6.0839 2.216e-03
#> BFP_Barcode_512115  -3.681  5.6765 2.646e-03
#> BFP_Barcode_48011    4.118  4.7879 3.379e-03
#> BFP_Barcode_98171   -4.555  6.6078 3.735e-03
#> BFP_Barcode_59254   -3.796  3.0752 4.254e-03
#> BFP_Barcode_161205   2.750  6.5719 4.324e-03
#> BFP_Barcode_42868    2.637  8.6485 4.979e-03
#> BFP_Barcode_229324   2.794  5.9315 5.184e-03
#> BFP_Barcode_77209   -3.044  3.4310 5.299e-03
#> BFP_Barcode_36539   -4.168  5.5932 5.372e-03
#> BFP_Barcode_129228   2.676  5.9773 6.011e-03
#> BFP_Barcode_2405     2.864  5.5492 6.193e-03
#> BFP_Barcode_88284    2.874  7.1320 6.352e-03
#> BFP_Barcode_168715  -2.912  4.3852 6.365e-03
#> BFP_Barcode_166107  -2.539  5.8003 7.316e-03
#> BFP_Barcode_91447    2.541  6.1674 7.745e-03
#> BFP_Barcode_2598    -2.650  6.1783 8.116e-03
#> BFP_Barcode_253178  -3.306  2.9260 9.127e-03
#> BFP_Barcode_391649  -3.356  2.4240 9.270e-03
#> BFP_Barcode_59600    4.028  5.2491 1.013e-02
#> BFP_Barcode_29222   -2.559  6.9335 1.077e-02
#> BFP_Barcode_354517   4.223  6.1892 1.098e-02
#> BFP_Barcode_187931   2.547  6.7514 1.135e-02
#> BFP_Barcode_113781   3.999  5.6149 1.159e-02
#> BFP_Barcode_158917  -2.409  5.1978 1.177e-02
#> BFP_Barcode_62483    3.178  2.1488 1.182e-02
#> BFP_Barcode_290418   2.986  5.5990 1.193e-02
#> BFP_Barcode_166522   2.643  6.6847 1.205e-02
#> BFP_Barcode_444801   3.318  3.4950 1.291e-02
#> BFP_Barcode_42402    3.567  4.8749 1.325e-02
#> BFP_Barcode_3081     2.575  5.4407 1.401e-02
#> BFP_Barcode_75816    2.634  5.8888 1.412e-02
#> BFP_Barcode_8715    -2.283  6.8440 1.415e-02
#> BFP_Barcode_257602   2.465  6.1124 1.441e-02
#> BFP_Barcode_416908  -5.567  0.7042 1.444e-02
#> BFP_Barcode_57889   -2.286  6.9700 1.461e-02
#> BFP_Barcode_500780   3.945  5.0105 1.479e-02
#> BFP_Barcode_227002  -3.757  5.8632 1.487e-02
#> BFP_Barcode_275847   2.702  6.4351 1.569e-02
#> BFP_Barcode_632     -2.887  4.5549 1.637e-02
#> BFP_Barcode_250328   2.345  4.8446 1.657e-02
#> BFP_Barcode_303423   3.655  3.5319 1.663e-02
#> BFP_Barcode_203500  -5.433  0.6250 1.734e-02
#> BFP_Barcode_30785    2.195  5.9821 1.845e-02
#> BFP_Barcode_14769   -2.612  2.9679 2.026e-02
#> BFP_Barcode_217320  -2.130  8.1873 2.115e-02
#> BFP_Barcode_592     -3.254  4.1906 2.119e-02
#> BFP_Barcode_26255    2.794  4.8022 2.294e-02
#> BFP_Barcode_105821   4.626  3.6404 2.313e-02
#> BFP_Barcode_401379   2.154  6.8773 2.397e-02
#> BFP_Barcode_125813  -2.982  4.9064 2.434e-02
#> BFP_Barcode_4264     3.333  5.0749 2.468e-02
#> BFP_Barcode_249263   4.713  3.1750 2.486e-02
#> BFP_Barcode_9578     2.165  4.7221 2.500e-02
#> BFP_Barcode_85597   -2.380  4.9704 2.755e-02
#> BFP_Barcode_71108    2.098  4.6546 3.070e-02
#> BFP_Barcode_282048  -3.683  3.2327 3.153e-02
#> BFP_Barcode_60720    3.495  5.7667 3.187e-02
#> BFP_Barcode_869      2.312  6.0717 3.325e-02
#> BFP_Barcode_348320   2.361  5.8559 3.479e-02
#> BFP_Barcode_509895  -2.287  5.4690 3.504e-02
#> BFP_Barcode_78261   -1.961  5.9398 3.625e-02
#> BFP_Barcode_8159    -1.931  6.9570 3.636e-02
#> BFP_Barcode_159570   1.909  9.5984 3.693e-02
#> BFP_Barcode_39998    3.054  4.4638 3.703e-02
#> BFP_Barcode_240647   2.467  6.3799 4.123e-02
#> BFP_Barcode_149173   1.896  5.8580 4.132e-02
#> BFP_Barcode_4850     2.152  5.0492 4.309e-02
#> BFP_Barcode_10834    1.958  4.0986 4.319e-02
#> BFP_Barcode_70225    2.308  3.7188 4.333e-02
#> BFP_Barcode_239117   3.594  4.7592 4.440e-02
#> BFP_Barcode_23613    2.785  3.9246 4.474e-02
#> BFP_Barcode_250048   1.999  6.3412 4.503e-02
#> BFP_Barcode_547197   2.442  5.0814 4.630e-02
#> BFP_Barcode_122618  -1.965  6.4383 4.686e-02
#> BFP_Barcode_264869   1.891  4.9514 4.804e-02
#> BFP_Barcode_196093  -2.567  3.2029 4.830e-02
#> BFP_Barcode_22565    1.884  7.2968 4.923e-02
#> BFP_Barcode_436560   2.106  6.2144 4.929e-02
#> BFP_Barcode_140952   1.894  6.4245 4.955e-02
#> BFP_Barcode_204604   3.145  5.6350 5.113e-02
#> BFP_Barcode_142719   2.984  5.4337 5.165e-02
#> BFP_Barcode_400391  -1.789 10.9677 5.285e-02
#> BFP_Barcode_14422    2.781  4.5149 5.350e-02
#> BFP_Barcode_1        1.847  6.3783 5.380e-02
#> BFP_Barcode_55675   -1.749  8.3255 5.429e-02
#> BFP_Barcode_268575   1.957  6.3320 5.552e-02
#> BFP_Barcode_39769   -2.570  2.9794 5.627e-02
#> BFP_Barcode_79222    1.957  5.6124 5.757e-02
#> BFP_Barcode_183226   2.842  4.8259 5.819e-02
#> BFP_Barcode_136913  -1.814  6.2126 5.828e-02
#> BFP_Barcode_59681   -2.859  5.0567 5.830e-02
#> BFP_Barcode_16820    1.882  4.6298 5.983e-02
#> BFP_Barcode_163332  -2.050  5.9295 6.051e-02
#> BFP_Barcode_64776    2.376  4.0914 6.367e-02
#> BFP_Barcode_228099   2.839  3.9847 6.370e-02
#> BFP_Barcode_19557   -2.049  7.0436 6.436e-02
#> BFP_Barcode_461589  -1.680  6.5909 6.549e-02
#> BFP_Barcode_82571   -2.591  2.8583 6.641e-02
#> BFP_Barcode_531673   1.717  7.6865 6.697e-02
#> BFP_Barcode_368637   1.727  5.8263 6.833e-02
#> BFP_Barcode_107824   2.223  3.9701 6.861e-02
#> BFP_Barcode_50204    1.776  6.6187 6.942e-02
#> BFP_Barcode_463726   2.995  2.8540 7.176e-02
#> BFP_Barcode_67071    2.558  4.2132 7.350e-02
#> BFP_Barcode_244391   1.639  7.3828 7.378e-02
#> BFP_Barcode_18218   -1.982  7.0071 7.486e-02
#> BFP_Barcode_85420   -2.146  4.7323 7.493e-02
#> BFP_Barcode_108364   2.343  4.9912 7.662e-02
#> BFP_Barcode_604439  -2.186  3.1321 7.748e-02
#> BFP_Barcode_195128   3.514  3.1693 7.765e-02
#> BFP_Barcode_23361   -1.796  3.9770 7.868e-02
#> BFP_Barcode_320540   1.910  4.9481 7.901e-02
#> BFP_Barcode_15652   -1.678  7.4382 8.002e-02
#> BFP_Barcode_9082     1.598  6.2394 8.038e-02
#> BFP_Barcode_12191   -1.613  8.6071 8.128e-02
#> BFP_Barcode_17825    1.655  7.7771 8.152e-02
#> BFP_Barcode_162261   1.670  5.2775 8.284e-02
#> BFP_Barcode_459273   2.741  4.7212 8.298e-02
#> BFP_Barcode_408103   2.269  4.4255 8.660e-02
#> BFP_Barcode_61170   -1.569  6.1556 8.681e-02
#> BFP_Barcode_238425  -2.673  4.9027 8.730e-02
#> BFP_Barcode_1870    -1.568  6.7719 8.765e-02
#> BFP_Barcode_174262   2.563  4.0907 8.856e-02
#> BFP_Barcode_64316   -2.534  4.8586 8.864e-02
#> BFP_Barcode_35459   -1.745  7.2936 9.082e-02
#> BFP_Barcode_211102  -1.661  5.9578 9.150e-02
#> BFP_Barcode_62514   -1.673  5.3364 9.322e-02
#> BFP_Barcode_247821   2.385  4.1843 9.664e-02
#> BFP_Barcode_302238  -1.618  8.3000 9.706e-02
#> BFP_Barcode_110402   1.603  6.7645 9.809e-02
#> BFP_Barcode_257906  -1.550  5.2872 9.872e-02
#> BFP_Barcode_129902  -1.573  6.7443 1.003e-01
#> BFP_Barcode_394311  -1.571  5.8541 1.017e-01
#> BFP_Barcode_10842   -2.475  5.0067 1.017e-01
#> BFP_Barcode_176655   1.487  6.9067 1.020e-01
#> BFP_Barcode_307250  -1.596  5.2801 1.028e-01
#> BFP_Barcode_29087   -1.535  4.3668 1.032e-01
#> BFP_Barcode_4165    -1.561  7.0755 1.045e-01
#> BFP_Barcode_245600  -1.473  7.5913 1.047e-01
#> BFP_Barcode_524885  -1.538  6.4859 1.048e-01
#> BFP_Barcode_154806  -1.469  7.2894 1.054e-01
#> BFP_Barcode_167860   1.730  5.8960 1.055e-01
#> BFP_Barcode_48084   -2.410  5.0960 1.056e-01
#> BFP_Barcode_29208   -1.461  8.3211 1.070e-01
#> BFP_Barcode_56785    1.564  5.9654 1.085e-01
#> BFP_Barcode_49462   -2.124  5.7587 1.087e-01
#> BFP_Barcode_56792    1.447  8.7449 1.088e-01
#> BFP_Barcode_462102   1.774  6.2194 1.089e-01
#> BFP_Barcode_426      1.555  5.5121 1.104e-01
#> BFP_Barcode_89937   -1.565  6.9727 1.120e-01
#> BFP_Barcode_9209     3.004  5.1939 1.131e-01
#> BFP_Barcode_201372  -1.431  8.4861 1.140e-01
#> BFP_Barcode_250044  -2.300  4.5588 1.149e-01
#> BFP_Barcode_223832   1.511  7.0017 1.152e-01
#> BFP_Barcode_217109   1.725  4.5659 1.155e-01
#> BFP_Barcode_796     -2.256  6.5832 1.164e-01
#> BFP_Barcode_102160   1.417 10.6073 1.167e-01
#> BFP_Barcode_567312   1.605  6.1334 1.171e-01
#> BFP_Barcode_235837  -1.515  4.5835 1.184e-01
#> BFP_Barcode_77131   -1.428  6.9417 1.186e-01
#> BFP_Barcode_2791     1.470  9.2522 1.188e-01
#> BFP_Barcode_116329  -1.489  5.7695 1.190e-01
#> BFP_Barcode_89363    2.157  2.3798 1.190e-01
#> BFP_Barcode_43272    1.854  4.2907 1.192e-01
#> BFP_Barcode_44967   -1.522  4.0338 1.193e-01
#> BFP_Barcode_10250    1.603  7.7527 1.194e-01
#> BFP_Barcode_416641   2.279  4.5381 1.197e-01
#> BFP_Barcode_498665  -1.465  6.3853 1.205e-01
#> BFP_Barcode_229036  -1.678  7.2316 1.207e-01
#> BFP_Barcode_154519  -1.420  8.7878 1.219e-01
#> BFP_Barcode_363818   1.588  4.8996 1.226e-01
#> BFP_Barcode_183691   1.393  7.8234 1.226e-01
#> BFP_Barcode_164207   2.522  6.1851 1.227e-01
#> BFP_Barcode_28077   -1.392  7.3505 1.230e-01
#> BFP_Barcode_26876    2.474  5.9706 1.236e-01
#> BFP_Barcode_32680   -1.406  9.5573 1.248e-01
#> BFP_Barcode_198370  -1.581  6.9734 1.267e-01
#> BFP_Barcode_420078  -2.321  5.3640 1.269e-01
#> BFP_Barcode_327828   1.385  7.7973 1.273e-01
#> BFP_Barcode_563071   1.417  6.0430 1.284e-01
#> BFP_Barcode_278655   1.361  8.3418 1.316e-01
#> BFP_Barcode_523643  -2.067  3.1322 1.321e-01
#> BFP_Barcode_117908   1.544  5.7353 1.331e-01
#> BFP_Barcode_5472     1.371  7.8761 1.332e-01
#> BFP_Barcode_342203   1.376  6.9375 1.340e-01
#> BFP_Barcode_48941    1.593  5.8601 1.364e-01
#> BFP_Barcode_52365    1.735  6.3879 1.365e-01
#> BFP_Barcode_320752  -1.360  9.1569 1.389e-01
#> BFP_Barcode_57631    1.382  5.4707 1.396e-01
#> BFP_Barcode_474982  -1.347  9.2924 1.397e-01
#> BFP_Barcode_291435   2.154  4.9866 1.401e-01
#> BFP_Barcode_21901   -1.591  5.2900 1.407e-01
#> BFP_Barcode_138779   2.027  5.5098 1.433e-01
#> BFP_Barcode_294271  -1.324  7.1442 1.436e-01
#> BFP_Barcode_440795   3.114  4.0377 1.450e-01
#> BFP_Barcode_125776  -1.329  7.1930 1.451e-01
#> BFP_Barcode_10229   -2.355  5.2209 1.461e-01
#> BFP_Barcode_342846   1.486  6.5017 1.462e-01
#> BFP_Barcode_231962  -1.305  8.0315 1.473e-01
#> BFP_Barcode_174955  -2.044  4.0139 1.511e-01
#> BFP_Barcode_388331  -1.312  6.4207 1.512e-01
#> BFP_Barcode_57747   -1.968  6.7733 1.528e-01
#> BFP_Barcode_148481   2.424  1.5304 1.531e-01
#>  [ reached 'max' / getOption("max.print") -- omitted 1095 rows ]