extractBarcodeReads

Identify and extract reads containing DNA barcodes from a FASTQ or BAM alignment file

Usage

extractBarcodeReads(
  infile,
  barcode_pattern = "[ACTG][ACTG][GC][AT][GC][ACTG][ACTG][AT][GC][AT]",
  constant = NULL,
  yieldSize = 1e+06,
  outfile = NULL,
  clean = TRUE,
  verbose = FALSE,
  overwrite = TRUE
)

Arguments

infile

Input fastq or bam file containing barcoded reads

barcode_pattern

Pattern to search for in the reads

constant

The constant region flanking the barcode, two constant regions can be input as a character vector. This should be set if the constant region is part of the barcode_pattern. Up to two constant regions can be set e.g. c("ACTG","TACG")

yieldSize

How many lines of input file to stream at each pass

outfile

Optional. path to save parsed reads as output file

clean

Logical. remove previous output files generated by extractBarcodeReads before current run

verbose

Logical. print detailed output to console

overwrite

Logical. Overwrite previous output files generated by extractBarcodeReads

Value

Returns a data-frame containing the read ID, the 10X cell ID, 10X umi, lintraceR DNA barcode sequence and lintraceR DNA barcode ID.

Examples

test.fq <- system.file("extdata", "test_extract.fastq.gz", package = "bartools")
extractBarcodeReads(infile = test.fq, barcode_pattern = "ACTG")
#> extractBarcodeReads
#> Input file:  /Library/Frameworks/R.framework/Versions/4.1/Resources/library/bartools/extdata/test_extract.fastq.gz
#> Format is FASTQ.
#> Extracting barcode reads.
#> Percentage of reads with barcode:  37.88
#> Parsing complete for: /Library/Frameworks/R.framework/Versions/4.1/Resources/library/bartools/extdata/test_extract.fastq.gz
#>                                                      Read.ID DNA.Barcode
#> 1    NB501056:187:HYMTFBGX7:1:11101:8116:1071 1:N:0:ATCACGAT        ACTG
#> 2    NB501056:187:HYMTFBGX7:1:11101:7788:1071 1:N:0:ATCACGAT        ACTG
#> 3    NB501056:187:HYMTFBGX7:1:11101:6349:1072 1:N:0:ATCACGAT        ACTG
#> 4   NB501056:187:HYMTFBGX7:1:11101:18392:1083 1:N:0:ATCACGAT        ACTG
#> 5   NB501056:187:HYMTFBGX7:1:11101:17447:1104 1:N:0:ATCACGAT        ACTG
#> 6    NB501056:187:HYMTFBGX7:1:11101:8404:1115 1:N:0:ATCACGAT        ACTG
#> 7   NB501056:187:HYMTFBGX7:1:11101:13246:1156 1:N:0:ATCACGAT        ACTG
#> 8   NB501056:187:HYMTFBGX7:1:11101:21490:1163 1:N:0:ATCACGAT        ACTG
#> 9   NB501056:187:HYMTFBGX7:1:11101:12278:1164 1:N:0:ATCACGAT        ACTG
#> 10   NB501056:187:HYMTFBGX7:1:11101:6685:1168 1:N:0:ATCACGAT        ACTG
#> 11   NB501056:187:HYMTFBGX7:1:11101:8164:1181 1:N:0:ATCACGAT        ACTG
#> 12   NB501056:187:HYMTFBGX7:1:11101:7849:1201 1:N:0:ATCACGAT        ACTG
#> 13   NB501056:187:HYMTFBGX7:1:11101:9563:1211 1:N:0:ATCACGAT        ACTG
#> 14   NB501056:187:HYMTFBGX7:1:11101:5323:1220 1:N:0:ATCACGAT        ACTG
#> 15  NB501056:187:HYMTFBGX7:1:11101:12753:1220 1:N:0:ATCACGAT        ACTG
#> 16  NB501056:187:HYMTFBGX7:1:11101:25104:1223 1:N:0:ATCACGAT        ACTG
#> 17   NB501056:187:HYMTFBGX7:1:11101:2625:1238 1:N:0:ATCACGAT        ACTG
#> 18  NB501056:187:HYMTFBGX7:1:11101:15862:1244 1:N:0:ATCACGAT        ACTG
#> 19  NB501056:187:HYMTFBGX7:1:11101:25834:1255 1:N:0:ATCACGAT        ACTG
#> 20   NB501056:187:HYMTFBGX7:1:11101:6120:1266 1:N:0:ATCACGAT        ACTG
#> 21  NB501056:187:HYMTFBGX7:1:11101:12743:1266 1:N:0:ATCACGAT        ACTG
#> 22   NB501056:187:HYMTFBGX7:1:11101:7790:1276 1:N:0:ATCACGAT        ACTG
#> 23   NB501056:187:HYMTFBGX7:1:11101:2571:1312 1:N:0:ATCACGAT        ACTG
#> 24  NB501056:187:HYMTFBGX7:1:11101:13173:1325 1:N:0:ATCACGAT        ACTG
#> 25  NB501056:187:HYMTFBGX7:1:11101:10280:1325 1:N:0:ATCACGAT        ACTG
#> 26  NB501056:187:HYMTFBGX7:1:11101:10014:1326 1:N:0:ATCACGAT        ACTG
#> 27  NB501056:187:HYMTFBGX7:1:11101:22198:1327 1:N:0:ATCACGAT        ACTG
#> 28  NB501056:187:HYMTFBGX7:1:11101:16263:1348 1:N:0:ATCACGAT        ACTG
#> 29   NB501056:187:HYMTFBGX7:1:11101:4766:1370 1:N:0:ATCACGAT        ACTG
#> 30   NB501056:187:HYMTFBGX7:1:11101:4979:1378 1:N:0:ATCACGAT        ACTG
#> 31  NB501056:187:HYMTFBGX7:1:11101:14816:1380 1:N:0:ATCACGAT        ACTG
#> 32  NB501056:187:HYMTFBGX7:1:11101:26516:1385 1:N:0:ATCACGAT        ACTG
#> 33   NB501056:187:HYMTFBGX7:1:11101:4325:1390 1:N:0:ATCACGAT        ACTG
#> 34  NB501056:187:HYMTFBGX7:1:11101:21205:1397 1:N:0:ATCACGAT        ACTG
#> 35   NB501056:187:HYMTFBGX7:1:11101:9324:1400 1:N:0:ATCACGAT        ACTG
#> 36   NB501056:187:HYMTFBGX7:1:11101:9265:1414 1:N:0:ATCACGAT        ACTG
#> 37  NB501056:187:HYMTFBGX7:1:11101:16118:1419 1:N:0:ATCACGAT        ACTG
#> 38  NB501056:187:HYMTFBGX7:1:11101:12375:1419 1:N:0:ATCACGAT        ACTG
#> 39  NB501056:187:HYMTFBGX7:1:11101:17157:1431 1:N:0:ATCACGAT        ACTG
#> 40  NB501056:187:HYMTFBGX7:1:11101:11904:1433 1:N:0:ATCACGAT        ACTG
#> 41   NB501056:187:HYMTFBGX7:1:11101:9607:1434 1:N:0:ATCACGAT        ACTG
#> 42  NB501056:187:HYMTFBGX7:1:11101:24060:1436 1:N:0:ATCACGAT        ACTG
#> 43   NB501056:187:HYMTFBGX7:1:11101:8760:1450 1:N:0:ATCACGAT        ACTG
#> 44  NB501056:187:HYMTFBGX7:1:11101:11711:1456 1:N:0:ATCACGAT        ACTG
#> 45  NB501056:187:HYMTFBGX7:1:11101:15172:1457 1:N:0:ATCACGAT        ACTG
#> 46  NB501056:187:HYMTFBGX7:1:11101:20206:1462 1:N:0:ATCACGAT        ACTG
#> 47  NB501056:187:HYMTFBGX7:1:11101:16777:1465 1:N:0:ATCACGAT        ACTG
#> 48   NB501056:187:HYMTFBGX7:1:11101:5737:1480 1:N:0:ATCACGAT        ACTG
#> 49   NB501056:187:HYMTFBGX7:1:11101:4922:1481 1:N:0:ATCACGAT        ACTG
#> 50  NB501056:187:HYMTFBGX7:1:11101:22221:1485 1:N:0:ATCACGAT        ACTG
#> 51   NB501056:187:HYMTFBGX7:1:11101:3053:1489 1:N:0:ATCACGAT        ACTG
#> 52   NB501056:187:HYMTFBGX7:1:11101:9324:1506 1:N:0:ATCACGAT        ACTG
#> 53   NB501056:187:HYMTFBGX7:1:11101:7189:1541 1:N:0:ATCACGAT        ACTG
#> 54  NB501056:187:HYMTFBGX7:1:11101:15957:1570 1:N:0:ATCACGAT        ACTG
#> 55  NB501056:187:HYMTFBGX7:1:11101:19328:1572 1:N:0:ATCACGAT        ACTG
#> 56   NB501056:187:HYMTFBGX7:1:11101:1817:1576 1:N:0:ATCACGAT        ACTG
#> 57   NB501056:187:HYMTFBGX7:1:11101:4763:1578 1:N:0:ATCACGAT        ACTG
#> 58  NB501056:187:HYMTFBGX7:1:11101:16058:1583 1:N:0:ATCACGAT        ACTG
#> 59  NB501056:187:HYMTFBGX7:1:11101:26545:1585 1:N:0:ATCACGAT        ACTG
#> 60   NB501056:187:HYMTFBGX7:1:11101:9432:1602 1:N:0:ATCACGAT        ACTG
#> 61  NB501056:187:HYMTFBGX7:1:11101:23141:1627 1:N:0:ATCACGAT        ACTG
#> 62  NB501056:187:HYMTFBGX7:1:11101:10327:1638 1:N:0:ATCACGAT        ACTG
#> 63  NB501056:187:HYMTFBGX7:1:11101:12573:1639 1:N:0:ATCACGAT        ACTG
#> 64  NB501056:187:HYMTFBGX7:1:11101:21927:1665 1:N:0:ATCACGAT        ACTG
#> 65  NB501056:187:HYMTFBGX7:1:11101:19420:1676 1:N:0:ATCACGAT        ACTG
#> 66  NB501056:187:HYMTFBGX7:1:11101:21769:1703 1:N:0:ATCACGAT        ACTG
#> 67   NB501056:187:HYMTFBGX7:1:11101:6388:1707 1:N:0:ATCACGAT        ACTG
#> 68   NB501056:187:HYMTFBGX7:1:11101:4421:1708 1:N:0:ATCACGAT        ACTG
#> 69  NB501056:187:HYMTFBGX7:1:11101:21922:1711 1:N:0:ATCACGAT        ACTG
#> 70  NB501056:187:HYMTFBGX7:1:11101:25717:1714 1:N:0:ATCACGAT        ACTG
#> 71  NB501056:187:HYMTFBGX7:1:11101:19603:1714 1:N:0:ATCACGAT        ACTG
#> 72   NB501056:187:HYMTFBGX7:1:11101:7456:1717 1:N:0:ATCACGAT        ACTG
#> 73  NB501056:187:HYMTFBGX7:1:11101:20504:1756 1:N:0:ATCACGAT        ACTG
#> 74  NB501056:187:HYMTFBGX7:1:11101:11096:1759 1:N:0:ATCACGAT        ACTG
#> 75   NB501056:187:HYMTFBGX7:1:11101:3899:1768 1:N:0:ATCACGAT        ACTG
#> 76  NB501056:187:HYMTFBGX7:1:11101:20586:1771 1:N:0:ATCACGAT        ACTG
#> 77  NB501056:187:HYMTFBGX7:1:11101:19545:1774 1:N:0:ATCACGAT        ACTG
#> 78   NB501056:187:HYMTFBGX7:1:11101:6186:1802 1:N:0:ATCACGAT        ACTG
#> 79   NB501056:187:HYMTFBGX7:1:11101:7172:1805 1:N:0:ATCACGAT        ACTG
#> 80   NB501056:187:HYMTFBGX7:1:11101:5233:1810 1:N:0:ATCACGAT        ACTG
#> 81  NB501056:187:HYMTFBGX7:1:11101:17758:1843 1:N:0:ATCACGAT        ACTG
#> 82   NB501056:187:HYMTFBGX7:1:11101:4450:1870 1:N:0:ATCACGAT        ACTG
#> 83  NB501056:187:HYMTFBGX7:1:11101:15477:1880 1:N:0:ATCACGAT        ACTG
#> 84  NB501056:187:HYMTFBGX7:1:11101:22665:1881 1:N:0:ATCACGAT        ACTG
#> 85   NB501056:187:HYMTFBGX7:1:11101:5353:1884 1:N:0:ATCACGAT        ACTG
#> 86  NB501056:187:HYMTFBGX7:1:11101:16111:1915 1:N:0:ATCACGAT        ACTG
#> 87  NB501056:187:HYMTFBGX7:1:11101:24479:1923 1:N:0:ATCACGAT        ACTG
#> 88  NB501056:187:HYMTFBGX7:1:11101:24221:1924 1:N:0:ATCACGAT        ACTG
#> 89  NB501056:187:HYMTFBGX7:1:11101:10490:1927 1:N:0:ATCACGAT        ACTG
#> 90  NB501056:187:HYMTFBGX7:1:11101:25277:1931 1:N:0:ATCACGAT        ACTG
#> 91   NB501056:187:HYMTFBGX7:1:11101:6864:1940 1:N:0:ATCACGAT        ACTG
#> 92   NB501056:187:HYMTFBGX7:1:11101:4116:1958 1:N:0:ATCACGAT        ACTG
#> 93  NB501056:187:HYMTFBGX7:1:11101:12885:1962 1:N:0:ATCACGAT        ACTG
#> 94  NB501056:187:HYMTFBGX7:1:11101:16050:1965 1:N:0:ATCACGAT        ACTG
#> 95  NB501056:187:HYMTFBGX7:1:11101:11504:1968 1:N:0:ATCACGAT        ACTG
#> 96  NB501056:187:HYMTFBGX7:1:11101:24340:1981 1:N:0:ATCACGAT        ACTG
#> 97  NB501056:187:HYMTFBGX7:1:11101:22076:1983 1:N:0:ATCACGAT        ACTG
#> 98  NB501056:187:HYMTFBGX7:1:11101:24686:2016 1:N:0:ATCACGAT        ACTG
#> 99   NB501056:187:HYMTFBGX7:1:11101:8780:2023 1:N:0:ATCACGAT        ACTG
#> 100  NB501056:187:HYMTFBGX7:1:11101:5354:2026 1:N:0:ATCACGAT        ACTG
#> 101  NB501056:187:HYMTFBGX7:1:11101:4370:2030 1:N:0:ATCACGAT        ACTG
#> 102  NB501056:187:HYMTFBGX7:1:11101:2873:2032 1:N:0:ATCACGAT        ACTG
#> 103 NB501056:187:HYMTFBGX7:1:11101:11484:2040 1:N:0:ATCACGAT        ACTG
#> 104  NB501056:187:HYMTFBGX7:1:11101:8881:2042 1:N:0:ATCACGAT        ACTG
#> 105 NB501056:187:HYMTFBGX7:1:11101:22607:2045 1:N:0:ATCACGAT        ACTG
#> 106 NB501056:187:HYMTFBGX7:1:11101:15885:2080 1:N:0:ATCACGAT        ACTG
#> 107 NB501056:187:HYMTFBGX7:1:11101:12723:2095 1:N:0:ATCACGAT        ACTG
#> 108 NB501056:187:HYMTFBGX7:1:11101:11761:2097 1:N:0:ATCACGAT        ACTG
#> 109  NB501056:187:HYMTFBGX7:1:11101:9123:2118 1:N:0:ATCACGAT        ACTG
#> 110 NB501056:187:HYMTFBGX7:1:11101:22896:2135 1:N:0:ATCACGAT        ACTG
#> 111  NB501056:187:HYMTFBGX7:1:11101:8436:2138 1:N:0:ATCACGAT        ACTG
#> 112 NB501056:187:HYMTFBGX7:1:11101:20040:2150 1:N:0:ATCACGAT        ACTG
#> 113  NB501056:187:HYMTFBGX7:1:11101:7447:2153 1:N:0:ATCACGAT        ACTG
#> 114 NB501056:187:HYMTFBGX7:1:11101:17627:2175 1:N:0:ATCACGAT        ACTG
#> 115 NB501056:187:HYMTFBGX7:1:11101:24022:2179 1:N:0:ATCACGAT        ACTG
#> 116  NB501056:187:HYMTFBGX7:1:11101:5726:2200 1:N:0:ATCACGAT        ACTG
#> 117 NB501056:187:HYMTFBGX7:1:11101:11808:2224 1:N:0:ATCACGAT        ACTG
#> 118 NB501056:187:HYMTFBGX7:1:11101:19441:2236 1:N:0:ATCACGAT        ACTG
#> 119 NB501056:187:HYMTFBGX7:1:11101:23270:2239 1:N:0:ATCACGAT        ACTG
#> 120 NB501056:187:HYMTFBGX7:1:11101:18055:2268 1:N:0:ATCACGAT        ACTG
#> 121 NB501056:187:HYMTFBGX7:1:11101:14998:2304 1:N:0:ATCACGAT        ACTG
#> 122 NB501056:187:HYMTFBGX7:1:11101:24785:2319 1:N:0:ATCACGAT        ACTG
#> 123  NB501056:187:HYMTFBGX7:1:11101:5534:2331 1:N:0:ATCACGAT        ACTG
#> 124  NB501056:187:HYMTFBGX7:1:11101:2134:2353 1:N:0:ATCACGAT        ACTG
#> 125 NB501056:187:HYMTFBGX7:1:11101:13881:2363 1:N:0:ATCACGAT        ACTG
#> 126  NB501056:187:HYMTFBGX7:1:11101:2609:2380 1:N:0:ATCACGAT        ACTG
#> 127 NB501056:187:HYMTFBGX7:1:11101:26404:2383 1:N:0:ATCACGAT        ACTG
#> 128  NB501056:187:HYMTFBGX7:1:11101:5986:2385 1:N:0:ATCACGAT        ACTG
#> 129 NB501056:187:HYMTFBGX7:1:11101:12603:2393 1:N:0:ATCACGAT        ACTG
#> 130  NB501056:187:HYMTFBGX7:1:11101:2122:2398 1:N:0:ATCACGAT        ACTG
#> 131  NB501056:187:HYMTFBGX7:1:11101:5507:2407 1:N:0:ATCACGAT        ACTG
#> 132 NB501056:187:HYMTFBGX7:1:11101:11555:2408 1:N:0:ATCACGAT        ACTG
#> 133 NB501056:187:HYMTFBGX7:1:11101:13509:2424 1:N:0:ATCACGAT        ACTG
#> 134  NB501056:187:HYMTFBGX7:1:11101:6788:2428 1:N:0:ATCACGAT        ACTG
#> 135  NB501056:187:HYMTFBGX7:1:11101:8602:2432 1:N:0:ATCACGAT        ACTG
#> 136 NB501056:187:HYMTFBGX7:1:11101:13736:2435 1:N:0:ATCACGAT        ACTG
#> 137 NB501056:187:HYMTFBGX7:1:11101:10425:2445 1:N:0:ATCACGAT        ACTG
#> 138 NB501056:187:HYMTFBGX7:1:11101:10892:2446 1:N:0:ATCACGAT        ACTG
#> 139 NB501056:187:HYMTFBGX7:1:11101:21967:2449 1:N:0:ATCACGAT        ACTG
#> 140 NB501056:187:HYMTFBGX7:1:11101:16445:2466 1:N:0:ATCACGAT        ACTG
#> 141  NB501056:187:HYMTFBGX7:1:11101:4203:2480 1:N:0:ATCACGAT        ACTG
#> 142 NB501056:187:HYMTFBGX7:1:11101:26021:2481 1:N:0:ATCACGAT        ACTG
#> 143  NB501056:187:HYMTFBGX7:1:11101:2590:2490 1:N:0:ATCACGAT        ACTG
#> 144 NB501056:187:HYMTFBGX7:1:11101:26535:2505 1:N:0:ATCACGAT        ACTG
#> 145 NB501056:187:HYMTFBGX7:1:11101:10126:2507 1:N:0:ATCACGAT        ACTG
#> 146 NB501056:187:HYMTFBGX7:1:11101:16351:2518 1:N:0:ATCACGAT        ACTG
#> 147 NB501056:187:HYMTFBGX7:1:11101:21372:2538 1:N:0:ATCACGAT        ACTG
#> 148 NB501056:187:HYMTFBGX7:1:11101:16757:2542 1:N:0:ATCACGAT        ACTG
#> 149 NB501056:187:HYMTFBGX7:1:11101:20845:2552 1:N:0:ATCACGAT        ACTG
#> 150 NB501056:187:HYMTFBGX7:1:11101:24880:2559 1:N:0:ATCACGAT        ACTG
#> 151 NB501056:187:HYMTFBGX7:1:11101:23221:2565 1:N:0:ATCACGAT        ACTG
#> 152  NB501056:187:HYMTFBGX7:1:11101:6835:2571 1:N:0:ATCACGAT        ACTG
#> 153 NB501056:187:HYMTFBGX7:1:11101:15114:2584 1:N:0:ATCACGAT        ACTG
#> 154  NB501056:187:HYMTFBGX7:1:11101:6425:2589 1:N:0:ATCACGAT        ACTG
#> 155 NB501056:187:HYMTFBGX7:1:11101:10265:2598 1:N:0:ATCACGAT        ACTG
#> 156 NB501056:187:HYMTFBGX7:1:11101:21005:2607 1:N:0:ATCACGAT        ACTG
#> 157 NB501056:187:HYMTFBGX7:1:11101:26268:2613 1:N:0:ATCACGAT        ACTG
#> 158  NB501056:187:HYMTFBGX7:1:11101:9581:2614 1:N:0:ATCACGAT        ACTG
#> 159  NB501056:187:HYMTFBGX7:1:11101:8738:2616 1:N:0:ATCACGAT        ACTG
#> 160 NB501056:187:HYMTFBGX7:1:11101:24507:2634 1:N:0:ATCACGAT        ACTG
#> 161  NB501056:187:HYMTFBGX7:1:11101:2601:2643 1:N:0:ATCACGAT        ACTG
#> 162 NB501056:187:HYMTFBGX7:1:11101:10480:2644 1:N:0:ATCACGAT        ACTG
#> 163 NB501056:187:HYMTFBGX7:1:11101:17790:2645 1:N:0:ATCACGAT        ACTG
#> 164 NB501056:187:HYMTFBGX7:1:11101:24475:2653 1:N:0:ATCACGAT        ACTG
#> 165  NB501056:187:HYMTFBGX7:1:11101:8985:2655 1:N:0:ATCACGAT        ACTG
#> 166 NB501056:187:HYMTFBGX7:1:11101:12010:2672 1:N:0:ATCACGAT        ACTG
#> 167 NB501056:187:HYMTFBGX7:1:11101:16973:2673 1:N:0:ATCACGAT        ACTG
#> 168 NB501056:187:HYMTFBGX7:1:11101:23368:2674 1:N:0:ATCACGAT        ACTG
#> 169 NB501056:187:HYMTFBGX7:1:11101:11533:2674 1:N:0:ATCACGAT        ACTG
#> 170  NB501056:187:HYMTFBGX7:1:11101:7616:2678 1:N:0:ATCACGAT        ACTG
#> 171 NB501056:187:HYMTFBGX7:1:11101:16614:2680 1:N:0:ATCACGAT        ACTG
#> 172 NB501056:187:HYMTFBGX7:1:11101:14312:2681 1:N:0:ATCACGAT        ACTG
#> 173 NB501056:187:HYMTFBGX7:1:11101:24159:2687 1:N:0:ATCACGAT        ACTG
#> 174 NB501056:187:HYMTFBGX7:1:11101:10940:2693 1:N:0:ATCACGAT        ACTG
#> 175 NB501056:187:HYMTFBGX7:1:11101:14313:2693 1:N:0:ATCACGAT        ACTG
#> 176 NB501056:187:HYMTFBGX7:1:11101:15033:2698 1:N:0:ATCACGAT        ACTG
#> 177 NB501056:187:HYMTFBGX7:1:11101:20538:2699 1:N:0:ATCACGAT        ACTG
#> 178 NB501056:187:HYMTFBGX7:1:11101:19248:2723 1:N:0:ATCACGAT        ACTG
#> 179  NB501056:187:HYMTFBGX7:1:11101:9972:2724 1:N:0:ATCACGAT        ACTG
#> 180 NB501056:187:HYMTFBGX7:1:11101:13799:2741 1:N:0:ATCACGAT        ACTG
#> 181  NB501056:187:HYMTFBGX7:1:11101:2974:2741 1:N:0:ATCACGAT        ACTG
#> 182 NB501056:187:HYMTFBGX7:1:11101:20517:2761 1:N:0:ATCACGAT        ACTG
#> 183 NB501056:187:HYMTFBGX7:1:11101:10892:2770 1:N:0:ATCACGAT        ACTG
#> 184  NB501056:187:HYMTFBGX7:1:11101:7388:2774 1:N:0:ATCACGAT        ACTG
#> 185  NB501056:187:HYMTFBGX7:1:11101:8859:2782 1:N:0:ATCACGAT        ACTG
#> 186  NB501056:187:HYMTFBGX7:1:11101:5777:2788 1:N:0:ATCACGAT        ACTG
#> 187 NB501056:187:HYMTFBGX7:1:11101:23507:2791 1:N:0:ATCACGAT        ACTG
#> 188  NB501056:187:HYMTFBGX7:1:11101:5414:2794 1:N:0:ATCACGAT        ACTG
#> 189 NB501056:187:HYMTFBGX7:1:11101:12241:2799 1:N:0:ATCACGAT        ACTG
#> 190 NB501056:187:HYMTFBGX7:1:11101:20322:2809 1:N:0:ATCACGAT        ACTG
#> 191  NB501056:187:HYMTFBGX7:1:11101:6665:2820 1:N:0:ATCACGAT        ACTG
#> 192 NB501056:187:HYMTFBGX7:1:11101:20941:2824 1:N:0:ATCACGAT        ACTG
#> 193  NB501056:187:HYMTFBGX7:1:11101:4955:2832 1:N:0:ATCACGAT        ACTG
#> 194  NB501056:187:HYMTFBGX7:1:11101:2575:2833 1:N:0:ATCACGAT        ACTG
#> 195 NB501056:187:HYMTFBGX7:1:11101:19246:2836 1:N:0:ATCACGAT        ACTG
#> 196 NB501056:187:HYMTFBGX7:1:11101:24949:2863 1:N:0:ATCACGAT        ACTG
#> 197  NB501056:187:HYMTFBGX7:1:11101:8121:2863 1:N:0:ATCACGAT        ACTG
#> 198 NB501056:187:HYMTFBGX7:1:11101:21012:2865 1:N:0:ATCACGAT        ACTG
#> 199 NB501056:187:HYMTFBGX7:1:11101:10614:2868 1:N:0:ATCACGAT        ACTG
#> 200  NB501056:187:HYMTFBGX7:1:11101:6922:2876 1:N:0:ATCACGAT        ACTG
#> 201  NB501056:187:HYMTFBGX7:1:11101:1911:2887 1:N:0:ATCACGAT        ACTG
#> 202 NB501056:187:HYMTFBGX7:1:11101:16922:2891 1:N:0:ATCACGAT        ACTG
#> 203 NB501056:187:HYMTFBGX7:1:11101:22107:2905 1:N:0:ATCACGAT        ACTG
#> 204 NB501056:187:HYMTFBGX7:1:11101:21109:2927 1:N:0:ATCACGAT        ACTG
#> 205  NB501056:187:HYMTFBGX7:1:11101:2863:2943 1:N:0:ATCACGAT        ACTG
#> 206 NB501056:187:HYMTFBGX7:1:11101:23675:2951 1:N:0:ATCACGAT        ACTG
#> 207  NB501056:187:HYMTFBGX7:1:11101:1673:2958 1:N:0:ATCACGAT        ACTG
#> 208 NB501056:187:HYMTFBGX7:1:11101:14816:2959 1:N:0:ATCACGAT        ACTG
#> 209 NB501056:187:HYMTFBGX7:1:11101:18758:2970 1:N:0:ATCACGAT        ACTG
#> 210 NB501056:187:HYMTFBGX7:1:11101:15659:2974 1:N:0:ATCACGAT        ACTG
#> 211 NB501056:187:HYMTFBGX7:1:11101:10918:2997 1:N:0:ATCACGAT        ACTG
#> 212 NB501056:187:HYMTFBGX7:1:11101:10233:2998 1:N:0:ATCACGAT        ACTG
#> 213 NB501056:187:HYMTFBGX7:1:11101:23966:3009 1:N:0:ATCACGAT        ACTG
#> 214  NB501056:187:HYMTFBGX7:1:11101:9895:3023 1:N:0:ATCACGAT        ACTG
#> 215 NB501056:187:HYMTFBGX7:1:11101:21004:3046 1:N:0:ATCACGAT        ACTG
#> 216 NB501056:187:HYMTFBGX7:1:11101:19330:3049 1:N:0:ATCACGAT        ACTG
#> 217  NB501056:187:HYMTFBGX7:1:11101:5428:3049 1:N:0:ATCACGAT        ACTG
#> 218  NB501056:187:HYMTFBGX7:1:11101:9225:3053 1:N:0:ATCACGAT        ACTG
#> 219 NB501056:187:HYMTFBGX7:1:11101:23850:3057 1:N:0:ATCACGAT        ACTG
#> 220 NB501056:187:HYMTFBGX7:1:11101:23615:3082 1:N:0:ATCACGAT        ACTG
#> 221 NB501056:187:HYMTFBGX7:1:11101:15453:3097 1:N:0:ATCACGAT        ACTG
#> 222  NB501056:187:HYMTFBGX7:1:11101:7655:3100 1:N:0:ATCACGAT        ACTG
#> 223  NB501056:187:HYMTFBGX7:1:11101:4966:3103 1:N:0:ATCACGAT        ACTG
#> 224  NB501056:187:HYMTFBGX7:1:11101:8641:3115 1:N:0:ATCACGAT        ACTG
#> 225  NB501056:187:HYMTFBGX7:1:11101:7296:3142 1:N:0:ATCACGAT        ACTG
#> 226 NB501056:187:HYMTFBGX7:1:11101:12659:3154 1:N:0:ATCACGAT        ACTG
#> 227 NB501056:187:HYMTFBGX7:1:11101:22169:3162 1:N:0:ATCACGAT        ACTG
#> 228  NB501056:187:HYMTFBGX7:1:11101:6911:3164 1:N:0:ATCACGAT        ACTG
#> 229  NB501056:187:HYMTFBGX7:1:11101:9029:3169 1:N:0:ATCACGAT        ACTG
#> 230  NB501056:187:HYMTFBGX7:1:11101:9475:3185 1:N:0:ATCACGAT        ACTG
#> 231  NB501056:187:HYMTFBGX7:1:11101:5069:3193 1:N:0:ATCACGAT        ACTG
#> 232 NB501056:187:HYMTFBGX7:1:11101:15375:3201 1:N:0:ATCACGAT        ACTG
#> 233  NB501056:187:HYMTFBGX7:1:11101:3299:3208 1:N:0:ATCACGAT        ACTG
#> 234  NB501056:187:HYMTFBGX7:1:11101:3412:3211 1:N:0:ATCACGAT        ACTG
#> 235  NB501056:187:HYMTFBGX7:1:11101:9979:3220 1:N:0:ATCACGAT        ACTG
#> 236 NB501056:187:HYMTFBGX7:1:11101:21712:3238 1:N:0:ATCACGAT        ACTG
#> 237  NB501056:187:HYMTFBGX7:1:11101:8451:3239 1:N:0:ATCACGAT        ACTG
#> 238 NB501056:187:HYMTFBGX7:1:11101:13517:3244 1:N:0:ATCACGAT        ACTG
#> 239 NB501056:187:HYMTFBGX7:1:11101:15571:3246 1:N:0:ATCACGAT        ACTG
#> 240  NB501056:187:HYMTFBGX7:1:11101:9253:3259 1:N:0:ATCACGAT        ACTG
#> 241 NB501056:187:HYMTFBGX7:1:11101:20112:3263 1:N:0:ATCACGAT        ACTG
#> 242 NB501056:187:HYMTFBGX7:1:11101:16020:3274 1:N:0:ATCACGAT        ACTG
#> 243 NB501056:187:HYMTFBGX7:1:11101:23265:3277 1:N:0:ATCACGAT        ACTG
#> 244 NB501056:187:HYMTFBGX7:1:11101:23068:3278 1:N:0:ATCACGAT        ACTG
#> 245 NB501056:187:HYMTFBGX7:1:11101:10504:3287 1:N:0:ATCACGAT        ACTG
#> 246  NB501056:187:HYMTFBGX7:1:11101:4958:3288 1:N:0:ATCACGAT        ACTG
#> 247  NB501056:187:HYMTFBGX7:1:11101:2766:3291 1:N:0:ATCACGAT        ACTG
#> 248  NB501056:187:HYMTFBGX7:1:11101:1991:3301 1:N:0:ATCACGAT        ACTG
#> 249 NB501056:187:HYMTFBGX7:1:11101:20220:3307 1:N:0:ATCACGAT        ACTG
#> 250 NB501056:187:HYMTFBGX7:1:11101:16024:3321 1:N:0:ATCACGAT        ACTG
#> 251 NB501056:187:HYMTFBGX7:1:11101:25612:3322 1:N:0:ATCACGAT        ACTG
#> 252  NB501056:187:HYMTFBGX7:1:11101:6320:3328 1:N:0:ATCACGAT        ACTG
#> 253 NB501056:187:HYMTFBGX7:1:11101:14661:3333 1:N:0:ATCACGAT        ACTG
#> 254 NB501056:187:HYMTFBGX7:1:11101:22796:3335 1:N:0:ATCACGAT        ACTG
#> 255  NB501056:187:HYMTFBGX7:1:11101:9217:3338 1:N:0:ATCACGAT        ACTG
#> 256 NB501056:187:HYMTFBGX7:1:11101:19247:3358 1:N:0:ATCACGAT        ACTG
#> 257 NB501056:187:HYMTFBGX7:1:11101:16359:3358 1:N:0:ATCACGAT        ACTG
#> 258 NB501056:187:HYMTFBGX7:1:11101:14993:3359 1:N:0:ATCACGAT        ACTG
#> 259  NB501056:187:HYMTFBGX7:1:11101:4307:3359 1:N:0:ATCACGAT        ACTG
#> 260  NB501056:187:HYMTFBGX7:1:11101:4407:3386 1:N:0:ATCACGAT        ACTG
#> 261 NB501056:187:HYMTFBGX7:1:11101:12022:3395 1:N:0:ATCACGAT        ACTG
#> 262 NB501056:187:HYMTFBGX7:1:11101:17691:3399 1:N:0:ATCACGAT        ACTG
#> 263 NB501056:187:HYMTFBGX7:1:11101:25170:3407 1:N:0:ATCACGAT        ACTG
#> 264  NB501056:187:HYMTFBGX7:1:11101:5596:3413 1:N:0:ATCACGAT        ACTG
#> 265 NB501056:187:HYMTFBGX7:1:11101:14428:3415 1:N:0:ATCACGAT        ACTG
#> 266  NB501056:187:HYMTFBGX7:1:11101:2967:3423 1:N:0:ATCACGAT        ACTG
#> 267  NB501056:187:HYMTFBGX7:1:11101:6074:3429 1:N:0:ATCACGAT        ACTG
#> 268 NB501056:187:HYMTFBGX7:1:11101:10883:3434 1:N:0:ATCACGAT        ACTG
#> 269  NB501056:187:HYMTFBGX7:1:11101:5075:3438 1:N:0:ATCACGAT        ACTG
#> 270 NB501056:187:HYMTFBGX7:1:11101:18056:3443 1:N:0:ATCACGAT        ACTG
#> 271 NB501056:187:HYMTFBGX7:1:11101:17747:3448 1:N:0:ATCACGAT        ACTG
#> 272 NB501056:187:HYMTFBGX7:1:11101:11932:3450 1:N:0:ATCACGAT        ACTG
#> 273 NB501056:187:HYMTFBGX7:1:11101:21156:3456 1:N:0:ATCACGAT        ACTG
#> 274 NB501056:187:HYMTFBGX7:1:11101:10902:3461 1:N:0:ATCACGAT        ACTG
#> 275  NB501056:187:HYMTFBGX7:1:11101:4424:3463 1:N:0:ATCACGAT        ACTG
#> 276 NB501056:187:HYMTFBGX7:1:11101:16116:3467 1:N:0:ATCACGAT        ACTG
#> 277 NB501056:187:HYMTFBGX7:1:11101:10029:3491 1:N:0:ATCACGAT        ACTG
#> 278 NB501056:187:HYMTFBGX7:1:11101:22122:3494 1:N:0:ATCACGAT        ACTG
#> 279 NB501056:187:HYMTFBGX7:1:11101:21357:3501 1:N:0:ATCACGAT        ACTG
#> 280 NB501056:187:HYMTFBGX7:1:11101:14972:3502 1:N:0:ATCACGAT        ACTG
#> 281 NB501056:187:HYMTFBGX7:1:11101:15108:3511 1:N:0:ATCACGAT        ACTG
#> 282  NB501056:187:HYMTFBGX7:1:11101:7612:3518 1:N:0:ATCACGAT        ACTG
#> 283  NB501056:187:HYMTFBGX7:1:11101:3999:3536 1:N:0:ATCACGAT        ACTG
#> 284 NB501056:187:HYMTFBGX7:1:11101:19777:3541 1:N:0:ATCACGAT        ACTG
#> 285 NB501056:187:HYMTFBGX7:1:11101:19974:3542 1:N:0:ATCACGAT        ACTG
#> 286  NB501056:187:HYMTFBGX7:1:11101:8664:3544 1:N:0:ATCACGAT        ACTG
#> 287 NB501056:187:HYMTFBGX7:1:11101:22080:3546 1:N:0:ATCACGAT        ACTG
#> 288 NB501056:187:HYMTFBGX7:1:11101:13999:3557 1:N:0:ATCACGAT        ACTG
#> 289 NB501056:187:HYMTFBGX7:1:11101:14298:3569 1:N:0:ATCACGAT        ACTG
#> 290  NB501056:187:HYMTFBGX7:1:11101:6814:3570 1:N:0:ATCACGAT        ACTG
#> 291  NB501056:187:HYMTFBGX7:1:11101:3982:3571 1:N:0:ATCACGAT        ACTG
#> 292 NB501056:187:HYMTFBGX7:1:11101:23996:3580 1:N:0:ATCACGAT        ACTG
#> 293 NB501056:187:HYMTFBGX7:1:11101:18925:3610 1:N:0:ATCACGAT        ACTG
#> 294 NB501056:187:HYMTFBGX7:1:11101:18703:3616 1:N:0:ATCACGAT        ACTG
#> 295  NB501056:187:HYMTFBGX7:1:11101:1250:3619 1:N:0:ATCACGAT        ACTG
#> 296 NB501056:187:HYMTFBGX7:1:11101:22703:3621 1:N:0:ATCACGAT        ACTG
#> 297  NB501056:187:HYMTFBGX7:1:11101:9290:3626 1:N:0:ATCACGAT        ACTG
#> 298 NB501056:187:HYMTFBGX7:1:11101:18250:3628 1:N:0:ATCACGAT        ACTG
#> 299 NB501056:187:HYMTFBGX7:1:11101:15343:3629 1:N:0:ATCACGAT        ACTG
#> 300 NB501056:187:HYMTFBGX7:1:11101:19959:3630 1:N:0:ATCACGAT        ACTG
#> 301 NB501056:187:HYMTFBGX7:1:11101:17144:3631 1:N:0:ATCACGAT        ACTG
#> 302 NB501056:187:HYMTFBGX7:1:11101:18666:3631 1:N:0:ATCACGAT        ACTG
#> 303  NB501056:187:HYMTFBGX7:1:11101:6501:3635 1:N:0:ATCACGAT        ACTG
#> 304 NB501056:187:HYMTFBGX7:1:11101:16169:3642 1:N:0:ATCACGAT        ACTG
#> 305 NB501056:187:HYMTFBGX7:1:11101:21204:3646 1:N:0:ATCACGAT        ACTG
#> 306 NB501056:187:HYMTFBGX7:1:11101:22846:3654 1:N:0:ATCACGAT        ACTG
#> 307  NB501056:187:HYMTFBGX7:1:11101:6809:3666 1:N:0:ATCACGAT        ACTG
#> 308  NB501056:187:HYMTFBGX7:1:11101:9534:3668 1:N:0:ATCACGAT        ACTG
#> 309  NB501056:187:HYMTFBGX7:1:11101:5396:3669 1:N:0:ATCACGAT        ACTG
#> 310 NB501056:187:HYMTFBGX7:1:11101:10599:3672 1:N:0:ATCACGAT        ACTG
#> 311  NB501056:187:HYMTFBGX7:1:11101:7014:3673 1:N:0:ATCACGAT        ACTG
#> 312  NB501056:187:HYMTFBGX7:1:11101:6912:3685 1:N:0:ATCACGAT        ACTG
#> 313 NB501056:187:HYMTFBGX7:1:11101:24144:3685 1:N:0:ATCACGAT        ACTG
#> 314 NB501056:187:HYMTFBGX7:1:11101:18284:3712 1:N:0:ATCACGAT        ACTG
#> 315 NB501056:187:HYMTFBGX7:1:11101:21271:3712 1:N:0:ATCACGAT        ACTG
#> 316 NB501056:187:HYMTFBGX7:1:11101:23721:3713 1:N:0:ATCACGAT        ACTG
#> 317 NB501056:187:HYMTFBGX7:1:11101:13342:3719 1:N:0:ATCACGAT        ACTG
#> 318 NB501056:187:HYMTFBGX7:1:11101:14356:3729 1:N:0:ATCACGAT        ACTG
#> 319  NB501056:187:HYMTFBGX7:1:11101:8514:3730 1:N:0:ATCACGAT        ACTG
#> 320 NB501056:187:HYMTFBGX7:1:11101:13499:3732 1:N:0:ATCACGAT        ACTG
#> 321  NB501056:187:HYMTFBGX7:1:11101:2707:3773 1:N:0:ATCACGAT        ACTG
#> 322 NB501056:187:HYMTFBGX7:1:11101:25922:3782 1:N:0:ATCACGAT        ACTG
#> 323  NB501056:187:HYMTFBGX7:1:11101:8702:3795 1:N:0:ATCACGAT        ACTG
#> 324 NB501056:187:HYMTFBGX7:1:11101:23601:3802 1:N:0:ATCACGAT        ACTG
#> 325  NB501056:187:HYMTFBGX7:1:11101:3038:3817 1:N:0:ATCACGAT        ACTG
#> 326  NB501056:187:HYMTFBGX7:1:11101:3973:3823 1:N:0:ATCACGAT        ACTG
#> 327 NB501056:187:HYMTFBGX7:1:11101:10107:3826 1:N:0:ATCACGAT        ACTG
#> 328  NB501056:187:HYMTFBGX7:1:11101:5724:3830 1:N:0:ATCACGAT        ACTG
#> 329  NB501056:187:HYMTFBGX7:1:11101:1133:3830 1:N:0:ATCACGAT        ACTG
#> 330 NB501056:187:HYMTFBGX7:1:11101:23560:3849 1:N:0:ATCACGAT        ACTG
#> 331 NB501056:187:HYMTFBGX7:1:11101:17708:3850 1:N:0:ATCACGAT        ACTG
#> 332 NB501056:187:HYMTFBGX7:1:11101:23302:3864 1:N:0:ATCACGAT        ACTG
#> 333  NB501056:187:HYMTFBGX7:1:11101:8154:3882 1:N:0:ATCACGAT        ACTG
#> 334 NB501056:187:HYMTFBGX7:1:11101:16821:3884 1:N:0:ATCACGAT        ACTG
#> 335 NB501056:187:HYMTFBGX7:1:11101:10974:3899 1:N:0:ATCACGAT        ACTG
#> 336 NB501056:187:HYMTFBGX7:1:11101:10069:3909 1:N:0:ATCACGAT        ACTG
#> 337 NB501056:187:HYMTFBGX7:1:11101:25775:3911 1:N:0:ATCACGAT        ACTG
#> 338  NB501056:187:HYMTFBGX7:1:11101:9416:3915 1:N:0:ATCACGAT        ACTG
#> 339  NB501056:187:HYMTFBGX7:1:11101:5117:3919 1:N:0:ATCACGAT        ACTG
#> 340 NB501056:187:HYMTFBGX7:1:11101:11228:3923 1:N:0:ATCACGAT        ACTG
#> 341 NB501056:187:HYMTFBGX7:1:11101:10625:3936 1:N:0:ATCACGAT        ACTG
#> 342  NB501056:187:HYMTFBGX7:1:11101:3328:3950 1:N:0:ATCACGAT        ACTG
#> 343  NB501056:187:HYMTFBGX7:1:11101:6399:3951 1:N:0:ATCACGAT        ACTG
#> 344 NB501056:187:HYMTFBGX7:1:11101:11452:3952 1:N:0:ATCACGAT        ACTG
#> 345 NB501056:187:HYMTFBGX7:1:11101:15802:3957 1:N:0:ATCACGAT        ACTG
#> 346 NB501056:187:HYMTFBGX7:1:11101:16341:3957 1:N:0:ATCACGAT        ACTG
#> 347 NB501056:187:HYMTFBGX7:1:11101:16279:3958 1:N:0:ATCACGAT        ACTG
#> 348  NB501056:187:HYMTFBGX7:1:11101:6958:3961 1:N:0:ATCACGAT        ACTG
#> 349 NB501056:187:HYMTFBGX7:1:11101:18602:3969 1:N:0:ATCACGAT        ACTG
#> 350 NB501056:187:HYMTFBGX7:1:11101:14934:3988 1:N:0:ATCACGAT        ACTG
#> 351  NB501056:187:HYMTFBGX7:1:11101:7581:3988 1:N:0:ATCACGAT        ACTG
#> 352  NB501056:187:HYMTFBGX7:1:11101:3064:3990 1:N:0:ATCACGAT        ACTG
#> 353 NB501056:187:HYMTFBGX7:1:11101:17456:3994 1:N:0:ATCACGAT        ACTG
#> 354 NB501056:187:HYMTFBGX7:1:11101:15707:4000 1:N:0:ATCACGAT        ACTG
#> 355 NB501056:187:HYMTFBGX7:1:11101:10531:4003 1:N:0:ATCACGAT        ACTG
#> 356 NB501056:187:HYMTFBGX7:1:11101:14620:4006 1:N:0:ATCACGAT        ACTG
#> 357  NB501056:187:HYMTFBGX7:1:11101:6306:4012 1:N:0:ATCACGAT        ACTG
#> 358 NB501056:187:HYMTFBGX7:1:11101:17069:4012 1:N:0:ATCACGAT        ACTG
#> 359 NB501056:187:HYMTFBGX7:1:11101:15850:4018 1:N:0:ATCACGAT        ACTG
#> 360  NB501056:187:HYMTFBGX7:1:11101:7786:4025 1:N:0:ATCACGAT        ACTG
#> 361 NB501056:187:HYMTFBGX7:1:11101:20508:4027 1:N:0:ATCACGAT        ACTG
#> 362 NB501056:187:HYMTFBGX7:1:11101:21036:4027 1:N:0:ATCACGAT        ACTG
#> 363 NB501056:187:HYMTFBGX7:1:11101:25861:4031 1:N:0:ATCACGAT        ACTG
#> 364 NB501056:187:HYMTFBGX7:1:11101:21647:4031 1:N:0:ATCACGAT        ACTG
#> 365 NB501056:187:HYMTFBGX7:1:11101:20561:4040 1:N:0:ATCACGAT        ACTG
#> 366 NB501056:187:HYMTFBGX7:1:11101:24091:4059 1:N:0:ATCACGAT        ACTG
#> 367 NB501056:187:HYMTFBGX7:1:11101:11602:4064 1:N:0:ATCACGAT        ACTG
#> 368  NB501056:187:HYMTFBGX7:1:11101:7316:4067 1:N:0:ATCACGAT        ACTG
#> 369  NB501056:187:HYMTFBGX7:1:11101:4478:4068 1:N:0:ATCACGAT        ACTG
#> 370  NB501056:187:HYMTFBGX7:1:11101:2575:4071 1:N:0:ATCACGAT        ACTG
#> 371  NB501056:187:HYMTFBGX7:1:11101:2444:4072 1:N:0:ATCACGAT        ACTG
#> 372 NB501056:187:HYMTFBGX7:1:11101:11504:4088 1:N:0:ATCACGAT        ACTG
#> 373 NB501056:187:HYMTFBGX7:1:11101:19273:4099 1:N:0:ATCACGAT        ACTG
#> 374  NB501056:187:HYMTFBGX7:1:11101:7203:4114 1:N:0:ATCACGAT        ACTG
#> 375 NB501056:187:HYMTFBGX7:1:11101:15117:4115 1:N:0:ATCACGAT        ACTG
#> 376  NB501056:187:HYMTFBGX7:1:11101:9255:4123 1:N:0:ATCACGAT        ACTG
#> 377 NB501056:187:HYMTFBGX7:1:11101:14607:4132 1:N:0:ATCACGAT        ACTG
#> 378 NB501056:187:HYMTFBGX7:1:11101:18536:4159 1:N:0:ATCACGAT        ACTG
#> 379  NB501056:187:HYMTFBGX7:1:11101:8873:4164 1:N:0:ATCACGAT        ACTG
#> 380 NB501056:187:HYMTFBGX7:1:11101:22354:4181 1:N:0:ATCACGAT        ACTG
#> 381  NB501056:187:HYMTFBGX7:1:11101:6984:4184 1:N:0:ATCACGAT        ACTG
#> 382 NB501056:187:HYMTFBGX7:1:11101:15177:4191 1:N:0:ATCACGAT        ACTG
#> 383 NB501056:187:HYMTFBGX7:1:11101:13619:4198 1:N:0:ATCACGAT        ACTG
#> 384 NB501056:187:HYMTFBGX7:1:11101:26313:4213 1:N:0:ATCACGAT        ACTG
#> 385 NB501056:187:HYMTFBGX7:1:11101:25046:4221 1:N:0:ATCACGAT        ACTG
#> 386 NB501056:187:HYMTFBGX7:1:11101:23640:4223 1:N:0:ATCACGAT        ACTG
#> 387 NB501056:187:HYMTFBGX7:1:11101:22361:4225 1:N:0:ATCACGAT        ACTG
#> 388 NB501056:187:HYMTFBGX7:1:11101:10437:4235 1:N:0:ATCACGAT        ACTG
#> 389 NB501056:187:HYMTFBGX7:1:11101:17665:4248 1:N:0:ATCACGAT        ACTG
#> 390 NB501056:187:HYMTFBGX7:1:11101:10530:4260 1:N:0:ATCACGAT        ACTG
#> 391 NB501056:187:HYMTFBGX7:1:11101:19556:4274 1:N:0:ATCACGAT        ACTG
#> 392  NB501056:187:HYMTFBGX7:1:11101:7847:4277 1:N:0:ATCACGAT        ACTG
#> 393 NB501056:187:HYMTFBGX7:1:11101:10676:4283 1:N:0:ATCACGAT        ACTG
#> 394 NB501056:187:HYMTFBGX7:1:11101:16325:4284 1:N:0:ATCACGAT        ACTG
#> 395  NB501056:187:HYMTFBGX7:1:11101:4321:4313 1:N:0:ATCACGAT        ACTG
#> 396 NB501056:187:HYMTFBGX7:1:11101:25917:4313 1:N:0:ATCACGAT        ACTG
#> 397  NB501056:187:HYMTFBGX7:1:11101:1342:4314 1:N:0:ATCACGAT        ACTG
#> 398 NB501056:187:HYMTFBGX7:1:11101:17500:4318 1:N:0:ATCACGAT        ACTG
#> 399  NB501056:187:HYMTFBGX7:1:11101:6055:4338 1:N:0:ATCACGAT        ACTG
#> 400 NB501056:187:HYMTFBGX7:1:11101:11700:4338 1:N:0:ATCACGAT        ACTG
#> 401 NB501056:187:HYMTFBGX7:1:11101:22335:4349 1:N:0:ATCACGAT        ACTG
#> 402 NB501056:187:HYMTFBGX7:1:11101:10629:4351 1:N:0:ATCACGAT        ACTG
#> 403 NB501056:187:HYMTFBGX7:1:11101:26219:4355 1:N:0:ATCACGAT        ACTG
#> 404 NB501056:187:HYMTFBGX7:1:11101:18611:4371 1:N:0:ATCACGAT        ACTG
#> 405 NB501056:187:HYMTFBGX7:1:11101:21067:4405 1:N:0:ATCACGAT        ACTG
#> 406 NB501056:187:HYMTFBGX7:1:11101:23153:4407 1:N:0:ATCACGAT        ACTG
#> 407 NB501056:187:HYMTFBGX7:1:11101:23603:4408 1:N:0:ATCACGAT        ACTG
#> 408 NB501056:187:HYMTFBGX7:1:11101:15707:4408 1:N:0:ATCACGAT        ACTG
#> 409  NB501056:187:HYMTFBGX7:1:11101:3311:4409 1:N:0:ATCACGAT        ACTG
#> 410 NB501056:187:HYMTFBGX7:1:11101:11678:4409 1:N:0:ATCACGAT        ACTG
#> 411  NB501056:187:HYMTFBGX7:1:11101:3272:4419 1:N:0:ATCACGAT        ACTG
#> 412 NB501056:187:HYMTFBGX7:1:11101:11125:4428 1:N:0:ATCACGAT        ACTG
#> 413  NB501056:187:HYMTFBGX7:1:11101:3162:4438 1:N:0:ATCACGAT        ACTG
#> 414 NB501056:187:HYMTFBGX7:1:11101:19791:4461 1:N:0:ATCACGAT        ACTG
#> 415 NB501056:187:HYMTFBGX7:1:11101:11120:4464 1:N:0:ATCACGAT        ACTG
#> 416  NB501056:187:HYMTFBGX7:1:11101:2115:4477 1:N:0:ATCACGAT        ACTG
#> 417 NB501056:187:HYMTFBGX7:1:11101:16679:4477 1:N:0:ATCACGAT        ACTG
#> 418 NB501056:187:HYMTFBGX7:1:11101:10708:4478 1:N:0:ATCACGAT        ACTG
#> 419 NB501056:187:HYMTFBGX7:1:11101:14559:4481 1:N:0:ATCACGAT        ACTG
#> 420 NB501056:187:HYMTFBGX7:1:11101:24874:4505 1:N:0:ATCACGAT        ACTG
#> 421 NB501056:187:HYMTFBGX7:1:11101:18710:4510 1:N:0:ATCACGAT        ACTG
#> 422  NB501056:187:HYMTFBGX7:1:11101:7752:4516 1:N:0:ATCACGAT        ACTG
#> 423 NB501056:187:HYMTFBGX7:1:11101:12373:4527 1:N:0:ATCACGAT        ACTG
#> 424 NB501056:187:HYMTFBGX7:1:11101:26717:4539 1:N:0:ATCACGAT        ACTG
#> 425 NB501056:187:HYMTFBGX7:1:11101:20685:4540 1:N:0:ATCACGAT        ACTG
#> 426 NB501056:187:HYMTFBGX7:1:11101:17134:4545 1:N:0:ATCACGAT        ACTG
#> 427 NB501056:187:HYMTFBGX7:1:11101:23332:4557 1:N:0:ATCACGAT        ACTG
#> 428 NB501056:187:HYMTFBGX7:1:11101:12322:4565 1:N:0:ATCACGAT        ACTG
#> 429  NB501056:187:HYMTFBGX7:1:11101:5275:4576 1:N:0:ATCACGAT        ACTG
#> 430  NB501056:187:HYMTFBGX7:1:11101:5485:4582 1:N:0:ATCACGAT        ACTG
#> 431 NB501056:187:HYMTFBGX7:1:11101:25149:4595 1:N:0:ATCACGAT        ACTG
#> 432 NB501056:187:HYMTFBGX7:1:11101:16034:4596 1:N:0:ATCACGAT        ACTG
#> 433 NB501056:187:HYMTFBGX7:1:11101:22096:4612 1:N:0:ATCACGAT        ACTG
#> 434 NB501056:187:HYMTFBGX7:1:11101:18498:4618 1:N:0:ATCACGAT        ACTG
#> 435 NB501056:187:HYMTFBGX7:1:11101:16664:4638 1:N:0:ATCACGAT        ACTG
#> 436  NB501056:187:HYMTFBGX7:1:11101:4186:4651 1:N:0:ATCACGAT        ACTG
#> 437 NB501056:187:HYMTFBGX7:1:11101:19450:4651 1:N:0:ATCACGAT        ACTG
#> 438 NB501056:187:HYMTFBGX7:1:11101:17312:4658 1:N:0:ATCACGAT        ACTG
#> 439 NB501056:187:HYMTFBGX7:1:11101:21976:4662 1:N:0:ATCACGAT        ACTG
#> 440 NB501056:187:HYMTFBGX7:1:11101:10294:4664 1:N:0:ATCACGAT        ACTG
#> 441  NB501056:187:HYMTFBGX7:1:11101:9303:4679 1:N:0:ATCACGAT        ACTG
#> 442  NB501056:187:HYMTFBGX7:1:11101:9750:4681 1:N:0:ATCACGAT        ACTG
#> 443 NB501056:187:HYMTFBGX7:1:11101:10679:4683 1:N:0:ATCACGAT        ACTG
#> 444 NB501056:187:HYMTFBGX7:1:11101:15600:4688 1:N:0:ATCACGAT        ACTG
#> 445 NB501056:187:HYMTFBGX7:1:11101:26122:4689 1:N:0:ATCACGAT        ACTG
#> 446 NB501056:187:HYMTFBGX7:1:11101:24458:4692 1:N:0:ATCACGAT        ACTG
#> 447 NB501056:187:HYMTFBGX7:1:11101:22393:4701 1:N:0:ATCACGAT        ACTG
#> 448  NB501056:187:HYMTFBGX7:1:11101:4931:4707 1:N:0:ATCACGAT        ACTG
#> 449  NB501056:187:HYMTFBGX7:1:11101:3692:4709 1:N:0:ATCACGAT        ACTG
#> 450 NB501056:187:HYMTFBGX7:1:11101:14984:4718 1:N:0:ATCACGAT        ACTG
#> 451 NB501056:187:HYMTFBGX7:1:11101:21823:4720 1:N:0:ATCACGAT        ACTG
#> 452 NB501056:187:HYMTFBGX7:1:11101:10628:4730 1:N:0:ATCACGAT        ACTG
#> 453 NB501056:187:HYMTFBGX7:1:11101:24338:4732 1:N:0:ATCACGAT        ACTG
#> 454 NB501056:187:HYMTFBGX7:1:11101:16700:4747 1:N:0:ATCACGAT        ACTG
#> 455 NB501056:187:HYMTFBGX7:1:11101:18243:4747 1:N:0:ATCACGAT        ACTG
#> 456 NB501056:187:HYMTFBGX7:1:11101:11762:4774 1:N:0:ATCACGAT        ACTG
#> 457 NB501056:187:HYMTFBGX7:1:11101:20747:4794 1:N:0:ATCACGAT        ACTG
#> 458 NB501056:187:HYMTFBGX7:1:11101:14763:4795 1:N:0:ATCACGAT        ACTG
#> 459 NB501056:187:HYMTFBGX7:1:11101:15403:4812 1:N:0:ATCACGAT        ACTG
#> 460 NB501056:187:HYMTFBGX7:1:11101:25141:4815 1:N:0:ATCACGAT        ACTG
#> 461 NB501056:187:HYMTFBGX7:1:11101:18285:4825 1:N:0:ATCACGAT        ACTG
#> 462  NB501056:187:HYMTFBGX7:1:11101:7828:4827 1:N:0:ATCACGAT        ACTG
#> 463  NB501056:187:HYMTFBGX7:1:11101:7201:4862 1:N:0:ATCACGAT        ACTG
#> 464  NB501056:187:HYMTFBGX7:1:11101:1515:4865 1:N:0:ATCACGAT        ACTG
#> 465  NB501056:187:HYMTFBGX7:1:11101:8357:4865 1:N:0:ATCACGAT        ACTG
#> 466 NB501056:187:HYMTFBGX7:1:11101:21537:4880 1:N:0:ATCACGAT        ACTG
#> 467 NB501056:187:HYMTFBGX7:1:11101:26760:4887 1:N:0:ATCACGAT        ACTG
#> 468  NB501056:187:HYMTFBGX7:1:11101:6661:4893 1:N:0:ATCACGAT        ACTG
#> 469  NB501056:187:HYMTFBGX7:1:11101:4087:4920 1:N:0:ATCACGAT        ACTG
#> 470 NB501056:187:HYMTFBGX7:1:11101:25669:4940 1:N:0:ATCACGAT        ACTG
#> 471  NB501056:187:HYMTFBGX7:1:11101:1938:4955 1:N:0:ATCACGAT        ACTG
#> 472 NB501056:187:HYMTFBGX7:1:11101:10342:4960 1:N:0:ATCACGAT        ACTG
#> 473  NB501056:187:HYMTFBGX7:1:11101:1379:4965 1:N:0:ATCACGAT        ACTG
#> 474 NB501056:187:HYMTFBGX7:1:11101:24791:4967 1:N:0:ATCACGAT        ACTG
#> 475 NB501056:187:HYMTFBGX7:1:11101:14043:4968 1:N:0:ATCACGAT        ACTG
#> 476 NB501056:187:HYMTFBGX7:1:11101:16250:4968 1:N:0:ATCACGAT        ACTG
#> 477  NB501056:187:HYMTFBGX7:1:11101:3865:4973 1:N:0:ATCACGAT        ACTG
#> 478  NB501056:187:HYMTFBGX7:1:11101:9519:5003 1:N:0:ATCACGAT        ACTG
#> 479 NB501056:187:HYMTFBGX7:1:11101:21071:5027 1:N:0:ATCACGAT        ACTG
#> 480  NB501056:187:HYMTFBGX7:1:11101:8265:5029 1:N:0:ATCACGAT        ACTG
#> 481  NB501056:187:HYMTFBGX7:1:11101:5441:5031 1:N:0:ATCACGAT        ACTG
#> 482 NB501056:187:HYMTFBGX7:1:11101:23322:5061 1:N:0:ATCACGAT        ACTG
#> 483  NB501056:187:HYMTFBGX7:1:11101:4616:5064 1:N:0:ATCACGAT        ACTG
#> 484 NB501056:187:HYMTFBGX7:1:11101:17336:5068 1:N:0:ATCACGAT        ACTG
#> 485 NB501056:187:HYMTFBGX7:1:11101:14092:5070 1:N:0:ATCACGAT        ACTG
#> 486 NB501056:187:HYMTFBGX7:1:11101:24313:5074 1:N:0:ATCACGAT        ACTG
#> 487 NB501056:187:HYMTFBGX7:1:11101:26381:5079 1:N:0:ATCACGAT        ACTG
#> 488 NB501056:187:HYMTFBGX7:1:11101:18188:5093 1:N:0:ATCACGAT        ACTG
#> 489 NB501056:187:HYMTFBGX7:1:11101:15919:5115 1:N:0:ATCACGAT        ACTG
#> 490  NB501056:187:HYMTFBGX7:1:11101:8536:5124 1:N:0:ATCACGAT        ACTG
#> 491 NB501056:187:HYMTFBGX7:1:11101:22081:5129 1:N:0:ATCACGAT        ACTG
#> 492 NB501056:187:HYMTFBGX7:1:11101:10140:5139 1:N:0:ATCACGAT        ACTG
#> 493 NB501056:187:HYMTFBGX7:1:11101:10424:5155 1:N:0:ATCACGAT        ACTG
#> 494  NB501056:187:HYMTFBGX7:1:11101:3921:5158 1:N:0:ATCACGAT        ACTG
#> 495 NB501056:187:HYMTFBGX7:1:11101:22154:5158 1:N:0:ATCACGAT        ACTG
#> 496  NB501056:187:HYMTFBGX7:1:11101:3528:5160 1:N:0:ATCACGAT        ACTG
#> 497 NB501056:187:HYMTFBGX7:1:11101:23984:5169 1:N:0:ATCACGAT        ACTG
#> 498 NB501056:187:HYMTFBGX7:1:11101:24741:5182 1:N:0:ATCACGAT        ACTG
#> 499  NB501056:187:HYMTFBGX7:1:11101:2791:5224 1:N:0:ATCACGAT        ACTG
#> 500 NB501056:187:HYMTFBGX7:1:11101:19133:5226 1:N:0:ATCACGAT        ACTG
#>  [ reached 'max' / getOption("max.print") -- omitted 447 rows ]
test.bam <- system.file("extdata", "test_extract_anno.bam", package = "bartools")
extractBarcodeReads(infile = test.bam, barcode_pattern = "ACTG")
#> extractBarcodeReads
#> Input file:  /Library/Frameworks/R.framework/Versions/4.1/Resources/library/bartools/extdata/test_extract_anno.bam
#> Format is BAM.
#> 
#> Parsing reads 0 to 1e+06
#> 76.8% of reads contain the barcode pattern
#> 64% of reads have a 10X cell ID match
#> 
#> Parsing reads 1e+06 to 2e+06
#> No reads contained a 10X cell ID. Skip
#> Identified 492 reads that contained the barcode pattern
#>                                     Read.ID
#>   1: A00130:99:HK225DSXX:1:1262:16351:17942
#>   2: A00130:99:HK225DSXX:1:1306:31512:33395
#>   3:  A00130:99:HK225DSXX:1:2556:7012:12195
#>   4:  A00130:99:HK225DSXX:1:2535:11939:4695
#>   5:  A00130:99:HK225DSXX:1:2435:10556:4366
#>  ---                                       
#> 488:  A00130:99:HK225DSXX:1:1463:1398:31125
#> 489: A00130:99:HK225DSXX:1:2514:31539:16595
#> 490:  A00130:99:HK225DSXX:1:2429:4237:29935
#> 491:  A00130:99:HK225DSXX:1:2275:28085:6762
#> 492:  A00130:99:HK225DSXX:1:2353:3884:21840
#>                                                                                                                                                    Read.Qual
#>   1: :FFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF,FFFFFFFFFFFFF:FFFFFFFFFFFFFFFFFFFFFFFFF
#>   2: FFFFFFFFFFFFFFFFF,FFFFFFF::FF:FFFFFFFFFFFFFFFFF:F:FFFFFFFFFFFFFFF:FFFFFF:F:FFFFFFFFF:FFFF:F:FFF:F:FFFFFFFFFFFFF:F,FFFFFFFFFFFFFFF:FF,:FFFFFFFFFFFFF,FF:
#>   3: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFFFFFFFFFFF:FFFFFFFFFF:FF:FFFF:FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF
#>   4: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFF:FFF:FFFFFFF:FFFFF:FFFFFFFFFFFFFFFFFFFFFFFFFFFFF
#>   5: FFFFFFFFFFFFFFFFFFFFFF,FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFFFFFF:FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFFFFF:FFFFFFFFFFFFFFFFFFFFFF
#>  ---                                                                                                                                                        
#> 488: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFF,FFFFFFFFFFFF:FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFFFFFFFFFFF:FF:F:FFFFFFFFF
#> 489:   FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFF,FFFFFF,FFFFFFFFFFFF:,FFFFFFFFFFF:FFFFFFFFFFFFF:FFFFFF,FFFFFFF:FFFFFFFFFFFFF:FFFF:FFFFFFF
#> 490: ::F,FFF:FF,,:F,:,F:FFF,,FF,,F::::F,,:,:F:,FF,FFF,F:F,FFFF:,F,FF,F,FFFFF:F,F:F::,F:FFF:F,FFF,FFF::F:,FF,FF,FF::F,:FF:F:F,F:FF:,:FFFF,FFFFFFFFF:FF:FFFFFF
#> 491: FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF::FFFFFFFFF:FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFFFFFF
#> 492: :FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF:FFFFF,FFFFFFFFFFFFFFFFFF:FFFFFFF,FFFFFF:FFFFF:
#>        Cell.10X.Barcode Read.10X.UMI DNA.Barcode
#>   1: AAACCCAAGACCAAGC-1 GGTCTCCATTAA        ACTG
#>   2: AAACCCAAGACCAAGC-1 GGTCTCCATTAA        ACTG
#>   3: AAACCCAAGACCAAGC-1 GGTCTCCATTAA        ACTG
#>   4: AAACCCAAGTCGCCCA-1 GGTTGGTACTGG        ACTG
#>   5: AAACCCAAGTTGCTCA-1 AGTGTTGCTCAC        ACTG
#>  ---                                            
#> 488: AACAACCCATTGGCAT-1 TACCACAGGCAT        ACTG
#> 489: AACAACCCATTGGCAT-1 TACCACAGGCAT        ACTG
#> 490: AACAACCGTACTCGAT-1 TTGTTCCGCATG        ACTG
#> 491: AACAACCTCCGACGGT-1 CATCCTAATGTC        ACTG
#> 492: AACAACCTCCGACGGT-1 CATCCTAATGTC        ACTG