Get dominant barcodes per sample

Takes a DGEList or dataframe of barcode counts, computes the percentage abundance of each barcode within each sample then returns a list of barcodes meeting a threshold abundance per sample

Usage

getDominantBarcodes(counts.obj, pct.thresh = 5)

Arguments

counts.obj

DGEList object or dataframe containing raw / normalised counts of barcodes

pct.thresh

percentage abundance threshold to call dominant barcodes

Value

Returns a named list containing vectors of dominant barcodes per sample

Details

getDominantBarcodes

get names of barcodes over a given percentage frequency per sample

Examples

data(test.dge)
getDominantBarcodes(test.dge, pct.thresh = 5)
#> $`T0-1`
#> [1] "BC_245016" "BC_121651"
#> 
#> $`T0-2`
#> [1] "BC_245016" "BC_121651"
#> 
#> $`S10-1`
#> [1] "BC_8419"   "BC_124796"
#> 
#> $`S10-2`
#> [1] "BC_8419"   "BC_124796"
#> 
#> $`S11-1`
#> [1] "BC_53234"  "BC_205581" "BC_90135"  "BC_172626" "BC_58978" 
#> 
#> $`S11-2`
#> [1] "BC_53234"  "BC_205581" "BC_90135"  "BC_172626" "BC_58978" 
#> 
#> $`S12-1`
#> [1] "BC_389078" "BC_159570" "BC_135438" "BC_500780" "BC_79755" 
#> 
#> $`S12-2`
#> [1] "BC_389078" "BC_159570" "BC_135438" "BC_500780" "BC_79755" 
#> 
#> $`S13-1`
#> [1] "BC_8419"   "BC_124796" "BC_102160" "BC_257382"
#> 
#> $`S13-2`
#> [1] "BC_8419"   "BC_124796" "BC_102160" "BC_257382"
#> 
#> $`S14-1`
#> [1] "BC_90135"  "BC_388103" "BC_31610" 
#> 
#> $`S14-2`
#> [1] "BC_90135"  "BC_388103" "BC_31610" 
#> 
#> $`S15-1`
#> [1] "BC_1478"   "BC_94579"  "BC_124796" "BC_8419"  
#> 
#> $`S15-2`
#> [1] "BC_1478"   "BC_94579"  "BC_124796" "BC_8419"  
#> 
#> $`S16-1`
#> [1] "BC_53234"  "BC_79755"  "BC_140320" "BC_31610"  "BC_248880" "BC_1478"  
#> 
#> $`S16-2`
#> [1] "BC_53234"  "BC_79755"  "BC_140320" "BC_31610"  "BC_248880" "BC_1478"  
#> 
#> $`S17-1`
#> [1] "BC_183691" "BC_90135"  "BC_400391"
#> 
#> $`S17-2`
#> [1] "BC_183691" "BC_90135"  "BC_400391"
#> 
#> $`S18-1`
#> [1] "BC_49397"  "BC_91412"  "BC_159570" "BC_62602"  "BC_93485"  "BC_70225" 
#> 
#> $`S18-2`
#> [1] "BC_49397"  "BC_91412"  "BC_159570" "BC_62602"  "BC_70225"  "BC_93485" 
#> 
#> $`S1-1`
#> [1] "BC_79755"  "BC_102160" "BC_59493"  "BC_23361"  "BC_53234"  "BC_400391"
#> [7] "BC_142112"
#> 
#> $`S1-2`
#> [1] "BC_79755"  "BC_102160" "BC_59493"  "BC_23361"  "BC_53234" 
#> 
#> $`S2-1`
#> [1] "BC_248880" "BC_65104"  "BC_9141"   "BC_1478"   "BC_31610"  "BC_79755" 
#> 
#> $`S2-2`
#> [1] "BC_248880" "BC_65104"  "BC_9141"   "BC_1478"   "BC_31610"  "BC_79755" 
#> 
#> $`S3-1`
#> [1] "BC_79755"  "BC_142024" "BC_102160"
#> 
#> $`S3-2`
#> [1] "BC_79755"  "BC_142024" "BC_102160" "BC_183691"
#> 
#> $`S4-1`
#> [1] "BC_53234"  "BC_58978"  "BC_62602"  "BC_68847"  "BC_140320" "BC_79755" 
#> 
#> $`S4-2`
#> [1] "BC_53234"  "BC_58978"  "BC_62602"  "BC_68847"  "BC_140320" "BC_79755" 
#> [7] "BC_327422"
#> 
#> $`S5-1`
#> [1] "BC_561471" "BC_118640" "BC_59493"  "BC_1607"  
#> 
#> $`S5-2`
#> [1] "BC_561471" "BC_118640" "BC_59493"  "BC_1607"   "BC_31610" 
#> 
#> $`S6-1`
#> [1] "BC_79755"  "BC_8419"   "BC_124796" "BC_4564"  
#> 
#> $`S6-2`
#> [1] "BC_79755"  "BC_8419"   "BC_124796" "BC_4564"  
#> 
#> $`S7-1`
#> [1] "BC_53234"  "BC_54442"  "BC_142024" "BC_140320" "BC_93485" 
#> 
#> $`S7-2`
#> [1] "BC_53234"  "BC_54442"  "BC_142024" "BC_140320" "BC_93485" 
#> 
#> $`S8-1`
#> [1] "BC_142024" "BC_90135"  "BC_58978"  "BC_93485" 
#> 
#> $`S8-2`
#> [1] "BC_142024" "BC_90135"  "BC_58978"  "BC_93485" 
#> 
#> $`S9-1`
#> [1] "BC_79755"  "BC_4198"   "BC_81579"  "BC_560839" "BC_118640" "BC_345889"
#> [7] "BC_400391"
#> 
#> $`S9-2`
#> [1] "BC_79755"  "BC_4198"   "BC_81579"  "BC_118640" "BC_560839" "BC_345889"
#> [7] "BC_400391"
#>